objc2_ui_kit

Struct NSLayoutManager

pub struct NSLayoutManager { /* private fields */ }

Apple’s documentation

Implementations

impl NSLayoutManager

pub fn init(this: Allocated<Self>) -> Retained<Self>

************************** Initialization ***************************

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

pub unsafe fn textStorage(&self) -> Option<Retained<NSTextStorage>>

************************* Text storage **************************

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn setTextStorage(&self, text_storage: Option<&NSTextStorage>)

Setter for textStorage.

Safety

This is unretained, you must ensure the object is kept alive while in use.

pub fn textContainers(&self) -> Retained<NSArray<NSTextContainer>>

************************** Text containers ***************************

pub fn addTextContainer(&self, container: &NSTextContainer)

pub fn insertTextContainer_atIndex( &self, container: &NSTextContainer, index: NSUInteger, )

pub fn removeTextContainerAtIndex(&self, index: NSUInteger)

pub fn textContainerChangedGeometry(&self, container: &NSTextContainer)

pub fn delegate( &self, ) -> Option<Retained<ProtocolObject<dyn NSLayoutManagerDelegate>>>

************************** Delegate ***************************

pub fn setDelegate( &self, delegate: Option<&ProtocolObject<dyn NSLayoutManagerDelegate>>, )

Setter for delegate.

This is a [weak property][objc2::topics::weak_property].

pub fn showsInvisibleCharacters(&self) -> bool

********************* Global layout manager options **********************

pub fn setShowsInvisibleCharacters(&self, shows_invisible_characters: bool)

Setter for showsInvisibleCharacters.

pub fn showsControlCharacters(&self) -> bool

pub fn setShowsControlCharacters(&self, shows_control_characters: bool)

Setter for showsControlCharacters.

pub fn usesFontLeading(&self) -> bool

pub fn setUsesFontLeading(&self, uses_font_leading: bool)

Setter for usesFontLeading.

pub fn allowsNonContiguousLayout(&self) -> bool

pub fn setAllowsNonContiguousLayout(&self, allows_non_contiguous_layout: bool)

Setter for allowsNonContiguousLayout.

pub fn hasNonContiguousLayout(&self) -> bool

pub fn limitsLayoutForSuspiciousContents(&self) -> bool

pub fn setLimitsLayoutForSuspiciousContents( &self, limits_layout_for_suspicious_contents: bool, )

Setter for limitsLayoutForSuspiciousContents.

pub fn usesDefaultHyphenation(&self) -> bool

pub fn setUsesDefaultHyphenation(&self, uses_default_hyphenation: bool)

Setter for usesDefaultHyphenation.

pub unsafe fn invalidateGlyphsForCharacterRange_changeInLength_actualCharacterRange( &self, char_range: NSRange, delta: NSInteger, actual_char_range: NSRangePointer, )

************************ Invalidation *************************

Safety

actual_char_range must be a valid pointer or null.

pub unsafe fn invalidateLayoutForCharacterRange_actualCharacterRange( &self, char_range: NSRange, actual_char_range: NSRangePointer, )

Safety

actual_char_range must be a valid pointer or null.

pub fn invalidateDisplayForCharacterRange(&self, char_range: NSRange)

pub fn invalidateDisplayForGlyphRange(&self, glyph_range: NSRange)

pub fn processEditingForTextStorage_edited_range_changeInLength_invalidatedRange( &self, text_storage: &NSTextStorage, edit_mask: NSTextStorageEditActions, new_char_range: NSRange, delta: NSInteger, invalidated_char_range: NSRange, )

pub fn ensureGlyphsForCharacterRange(&self, char_range: NSRange)

********************** Causing glyph generation and layout ***********************

pub fn ensureGlyphsForGlyphRange(&self, glyph_range: NSRange)

pub fn ensureLayoutForCharacterRange(&self, char_range: NSRange)

pub fn ensureLayoutForGlyphRange(&self, glyph_range: NSRange)

pub fn ensureLayoutForTextContainer(&self, container: &NSTextContainer)

pub fn ensureLayoutForBoundingRect_inTextContainer( &self, bounds: CGRect, container: &NSTextContainer, )

pub unsafe fn setGlyphs_properties_characterIndexes_font_forGlyphRange( &self, glyphs: NonNull<CGGlyph>, props: NonNull<NSGlyphProperty>, char_indexes: NonNull<NSUInteger>, a_font: &UIFont, glyph_range: NSRange, )

********************** Set glyphs and glyph properties ***********************

Safety

glyphs must be a valid pointer.

props must be a valid pointer.

char_indexes must be a valid pointer.

pub fn numberOfGlyphs(&self) -> NSUInteger

********************** Get glyphs and glyph properties ***********************

pub unsafe fn CGGlyphAtIndex_isValidIndex( &self, glyph_index: NSUInteger, is_valid_index: *mut Bool, ) -> CGGlyph

Safety

is_valid_index must be a valid pointer or null.

pub fn CGGlyphAtIndex(&self, glyph_index: NSUInteger) -> CGGlyph

pub fn isValidGlyphIndex(&self, glyph_index: NSUInteger) -> bool

pub fn propertyForGlyphAtIndex( &self, glyph_index: NSUInteger, ) -> NSGlyphProperty

pub fn characterIndexForGlyphAtIndex( &self, glyph_index: NSUInteger, ) -> NSUInteger

pub fn glyphIndexForCharacterAtIndex( &self, char_index: NSUInteger, ) -> NSUInteger

pub unsafe fn getGlyphsInRange_glyphs_properties_characterIndexes_bidiLevels( &self, glyph_range: NSRange, glyph_buffer: *mut CGGlyph, props: *mut NSGlyphProperty, char_index_buffer: *mut NSUInteger, bidi_level_buffer: *mut c_uchar, ) -> NSUInteger

Safety

glyph_buffer must be a valid pointer or null.

props must be a valid pointer or null.

char_index_buffer must be a valid pointer or null.

bidi_level_buffer must be a valid pointer or null.

pub fn setTextContainer_forGlyphRange( &self, container: &NSTextContainer, glyph_range: NSRange, )

pub fn setLineFragmentRect_forGlyphRange_usedRect( &self, fragment_rect: CGRect, glyph_range: NSRange, used_rect: CGRect, )

pub fn setExtraLineFragmentRect_usedRect_textContainer( &self, fragment_rect: CGRect, used_rect: CGRect, container: &NSTextContainer, )

pub fn setLocation_forStartOfGlyphRange( &self, location: CGPoint, glyph_range: NSRange, )

pub fn setNotShownAttribute_forGlyphAtIndex( &self, flag: bool, glyph_index: NSUInteger, )

pub fn setDrawsOutsideLineFragment_forGlyphAtIndex( &self, flag: bool, glyph_index: NSUInteger, )

pub fn setAttachmentSize_forGlyphRange( &self, attachment_size: CGSize, glyph_range: NSRange, )

pub unsafe fn getFirstUnlaidCharacterIndex_glyphIndex( &self, char_index: *mut NSUInteger, glyph_index: *mut NSUInteger, )

********************** Get layout information ***********************

Safety

char_index must be a valid pointer or null.

glyph_index must be a valid pointer or null.

pub fn firstUnlaidCharacterIndex(&self) -> NSUInteger

pub fn firstUnlaidGlyphIndex(&self) -> NSUInteger

pub unsafe fn textContainerForGlyphAtIndex_effectiveRange( &self, glyph_index: NSUInteger, effective_glyph_range: NSRangePointer, ) -> Option<Retained<NSTextContainer>>

Safety

effective_glyph_range must be a valid pointer or null.

pub unsafe fn textContainerForGlyphAtIndex_effectiveRange_withoutAdditionalLayout( &self, glyph_index: NSUInteger, effective_glyph_range: NSRangePointer, flag: bool, ) -> Option<Retained<NSTextContainer>>

Safety

effective_glyph_range must be a valid pointer or null.

pub fn usedRectForTextContainer(&self, container: &NSTextContainer) -> CGRect

pub unsafe fn lineFragmentRectForGlyphAtIndex_effectiveRange( &self, glyph_index: NSUInteger, effective_glyph_range: NSRangePointer, ) -> CGRect

Safety

effective_glyph_range must be a valid pointer or null.

pub unsafe fn lineFragmentRectForGlyphAtIndex_effectiveRange_withoutAdditionalLayout( &self, glyph_index: NSUInteger, effective_glyph_range: NSRangePointer, flag: bool, ) -> CGRect

Safety

effective_glyph_range must be a valid pointer or null.

pub unsafe fn lineFragmentUsedRectForGlyphAtIndex_effectiveRange( &self, glyph_index: NSUInteger, effective_glyph_range: NSRangePointer, ) -> CGRect

Safety

effective_glyph_range must be a valid pointer or null.

pub unsafe fn lineFragmentUsedRectForGlyphAtIndex_effectiveRange_withoutAdditionalLayout( &self, glyph_index: NSUInteger, effective_glyph_range: NSRangePointer, flag: bool, ) -> CGRect

Safety

effective_glyph_range must be a valid pointer or null.

pub fn extraLineFragmentRect(&self) -> CGRect

pub fn extraLineFragmentUsedRect(&self) -> CGRect

pub fn extraLineFragmentTextContainer( &self, ) -> Option<Retained<NSTextContainer>>

pub fn locationForGlyphAtIndex(&self, glyph_index: NSUInteger) -> CGPoint

pub fn notShownAttributeForGlyphAtIndex(&self, glyph_index: NSUInteger) -> bool

pub fn drawsOutsideLineFragmentForGlyphAtIndex( &self, glyph_index: NSUInteger, ) -> bool

pub fn attachmentSizeForGlyphAtIndex(&self, glyph_index: NSUInteger) -> CGSize

pub fn truncatedGlyphRangeInLineFragmentForGlyphAtIndex( &self, glyph_index: NSUInteger, ) -> NSRange

pub unsafe fn glyphRangeForCharacterRange_actualCharacterRange( &self, char_range: NSRange, actual_char_range: NSRangePointer, ) -> NSRange

********************** More sophisticated queries ***********************

Safety

actual_char_range must be a valid pointer or null.

pub unsafe fn characterRangeForGlyphRange_actualGlyphRange( &self, glyph_range: NSRange, actual_glyph_range: NSRangePointer, ) -> NSRange

Safety

actual_glyph_range must be a valid pointer or null.

pub fn glyphRangeForTextContainer(&self, container: &NSTextContainer) -> NSRange

pub fn rangeOfNominallySpacedGlyphsContainingIndex( &self, glyph_index: NSUInteger, ) -> NSRange

pub fn boundingRectForGlyphRange_inTextContainer( &self, glyph_range: NSRange, container: &NSTextContainer, ) -> CGRect

pub fn glyphRangeForBoundingRect_inTextContainer( &self, bounds: CGRect, container: &NSTextContainer, ) -> NSRange

pub fn glyphRangeForBoundingRectWithoutAdditionalLayout_inTextContainer( &self, bounds: CGRect, container: &NSTextContainer, ) -> NSRange

pub unsafe fn glyphIndexForPoint_inTextContainer_fractionOfDistanceThroughGlyph( &self, point: CGPoint, container: &NSTextContainer, partial_fraction: *mut CGFloat, ) -> NSUInteger

Safety

partial_fraction must be a valid pointer or null.

pub fn glyphIndexForPoint_inTextContainer( &self, point: CGPoint, container: &NSTextContainer, ) -> NSUInteger

pub fn fractionOfDistanceThroughGlyphForPoint_inTextContainer( &self, point: CGPoint, container: &NSTextContainer, ) -> CGFloat

pub unsafe fn characterIndexForPoint_inTextContainer_fractionOfDistanceBetweenInsertionPoints( &self, point: CGPoint, container: &NSTextContainer, partial_fraction: *mut CGFloat, ) -> NSUInteger

Safety

partial_fraction must be a valid pointer or null.

pub unsafe fn getLineFragmentInsertionPointsForCharacterAtIndex_alternatePositions_inDisplayOrder_positions_characterIndexes( &self, char_index: NSUInteger, a_flag: bool, d_flag: bool, positions: *mut CGFloat, char_indexes: *mut NSUInteger, ) -> NSUInteger

Safety

positions must be a valid pointer or null.

char_indexes must be a valid pointer or null.

pub fn enumerateLineFragmentsForGlyphRange_usingBlock( &self, glyph_range: NSRange, block: &DynBlock<dyn Fn(CGRect, CGRect, NonNull<NSTextContainer>, NSRange, NonNull<Bool>)>, )

pub fn enumerateEnclosingRectsForGlyphRange_withinSelectedGlyphRange_inTextContainer_usingBlock( &self, glyph_range: NSRange, selected_range: NSRange, text_container: &NSTextContainer, block: &DynBlock<dyn Fn(CGRect, NonNull<Bool>)>, )

pub fn drawBackgroundForGlyphRange_atPoint( &self, glyphs_to_show: NSRange, origin: CGPoint, )

********************** Drawing support ***********************

pub fn drawGlyphsForGlyphRange_atPoint( &self, glyphs_to_show: NSRange, origin: CGPoint, )

pub unsafe fn showCGGlyphs_positions_count_font_textMatrix_attributes_inContext( &self, glyphs: NonNull<CGGlyph>, positions: NonNull<CGPoint>, glyph_count: NSInteger, font: &UIFont, text_matrix: CGAffineTransform, attributes: &NSDictionary<NSAttributedStringKey, AnyObject>, cg_context: &CGContext, )

Safety

glyphs must be a valid pointer.

positions must be a valid pointer.

attributes generic should be of the correct type.

pub unsafe fn fillBackgroundRectArray_count_forCharacterRange_color( &self, rect_array: NonNull<CGRect>, rect_count: NSUInteger, char_range: NSRange, color: &UIColor, )

Safety

rect_array must be a valid pointer.

pub fn drawUnderlineForGlyphRange_underlineType_baselineOffset_lineFragmentRect_lineFragmentGlyphRange_containerOrigin( &self, glyph_range: NSRange, underline_val: NSUnderlineStyle, baseline_offset: CGFloat, line_rect: CGRect, line_glyph_range: NSRange, container_origin: CGPoint, )

pub fn underlineGlyphRange_underlineType_lineFragmentRect_lineFragmentGlyphRange_containerOrigin( &self, glyph_range: NSRange, underline_val: NSUnderlineStyle, line_rect: CGRect, line_glyph_range: NSRange, container_origin: CGPoint, )

pub fn drawStrikethroughForGlyphRange_strikethroughType_baselineOffset_lineFragmentRect_lineFragmentGlyphRange_containerOrigin( &self, glyph_range: NSRange, strikethrough_val: NSUnderlineStyle, baseline_offset: CGFloat, line_rect: CGRect, line_glyph_range: NSRange, container_origin: CGPoint, )

pub fn strikethroughGlyphRange_strikethroughType_lineFragmentRect_lineFragmentGlyphRange_containerOrigin( &self, glyph_range: NSRange, strikethrough_val: NSUnderlineStyle, line_rect: CGRect, line_glyph_range: NSRange, container_origin: CGPoint, )

impl NSLayoutManager

Methods declared on superclass NSObject.

pub fn new() -> Retained<Self>

impl NSLayoutManager

NSLayoutManagerDeprecated.

pub unsafe fn glyphAtIndex_isValidIndex( &self, glyph_index: NSUInteger, is_valid_index: *mut Bool, ) -> CGGlyph

Safety

is_valid_index must be a valid pointer or null.

pub fn glyphAtIndex(&self, glyph_index: NSUInteger) -> CGGlyph

pub fn hyphenationFactor(&self) -> CGFloat

👎Deprecated: Please use usesDefaultHyphenation or -[NSParagraphStyle hyphenationFactor] instead.

pub fn setHyphenationFactor(&self, hyphenation_factor: CGFloat)

👎Deprecated: Please use usesDefaultHyphenation or -[NSParagraphStyle hyphenationFactor] instead.

Setter for hyphenationFactor.

pub unsafe fn showCGGlyphs_positions_count_font_matrix_attributes_inContext( &self, glyphs: NonNull<CGGlyph>, positions: NonNull<CGPoint>, glyph_count: NSUInteger, font: &UIFont, text_matrix: CGAffineTransform, attributes: &NSDictionary<NSAttributedStringKey, AnyObject>, graphics_context: &CGContext, )

👎Deprecated

Safety

glyphs must be a valid pointer.

positions must be a valid pointer.

attributes generic should be of the correct type.