objc2_ui_kit

Struct NSTextRange

pub struct NSTextRange { /* private fields */ }

A class that represents a contiguous range between two locations inside document contents.

An NSTextRange consists of the starting and terminating locations. There the two basic properties: location and endLocation, respectively. The terminating location, endLocation, is directly following the last location in the range. For example, a location contains a range if (range.location < = location) & & (location < range.endLocation) is true.

See also Apple’s documentation

Implementations

impl NSTextRange

pub fn initWithLocation_endLocation( this: Allocated<Self>, location: &ProtocolObject<dyn NSTextLocation>, end_location: Option<&ProtocolObject<dyn NSTextLocation>>, ) -> Option<Retained<Self>>

Creates a new text range with the starting and ending locations you specify.

Parameters:

location: The starting location.

endLocation: The ending location.

pub fn initWithLocation( this: Allocated<Self>, location: &ProtocolObject<dyn NSTextLocation>, ) -> Retained<Self>

Creates a new text range at the location you specify.

Parameters:

location: An NSTextLocation.

pub unsafe fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn new() -> Retained<Self>

pub fn isEmpty(&self) -> bool

Returns whether the text range is empty.

pub fn location(&self) -> Retained<ProtocolObject<dyn NSTextLocation>>

The starting location of the text range.

pub fn endLocation(&self) -> Retained<ProtocolObject<dyn NSTextLocation>>

The ending location of the text range.

pub fn isEqualToTextRange(&self, text_range: &NSTextRange) -> bool

Compares two text ranges.

Parameters:

textRange: The range used to compare against the current range to evaluate for differences.

Returns: Returns true if the ranges are equal.

pub fn containsLocation( &self, location: &ProtocolObject<dyn NSTextLocation>, ) -> bool

Determines if the text location you specify is in the current text range.

Parameters:

location: An NSTextLocation.

Returns: Returns true if the location is in the range otherwise false .

pub fn containsRange(&self, text_range: &NSTextRange) -> bool

Determines if the text range you specify is in the current text range.

Parameters:

textRange: An NSTextRange.

Returns: Returns true if the range you provide is in the current range; otherwise false.

pub fn intersectsWithTextRange(&self, text_range: &NSTextRange) -> bool

Determines if two ranges intersect.

Parameters:

textRange: The range used to compare against the current range to evaluate for differences.

Returns: Returns true if the ranges intersect.

pub fn textRangeByIntersectingWithTextRange( &self, text_range: &NSTextRange, ) -> Option<Retained<Self>>

Returns the range, if any, where two text ranges intersect.

Parameters:

textRange: The range used to compare against the current range to evaluate for differences.

Returns: An <doc ://com.apple.documentation/documentation/foundation/nsrange> that represents the intersection of the ranges, or nil if they don’t intersect.

pub fn textRangeByFormingUnionWithTextRange( &self, text_range: &NSTextRange, ) -> Retained<Self>

Returns a new text range by forming the union with the text range you provide.

Parameters:

textRange: The range to use to create the union.

Returns: An NSTextRange that represent the union of the two ranges.