objc2_ui_kit

Struct UIBarButtonItem

pub struct UIBarButtonItem { /* private fields */ }

Apple’s documentation

Implementations

impl UIBarButtonItem

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

pub unsafe fn initWithImage_style_target_action( this: Allocated<Self>, image: Option<&UIImage>, style: UIBarButtonItemStyle, target: Option<&AnyObject>, action: Option<Sel>, ) -> Retained<Self>

Safety

target should be of the correct type.

action must be a valid selector.

pub unsafe fn initWithImage_landscapeImagePhone_style_target_action( this: Allocated<Self>, image: Option<&UIImage>, landscape_image_phone: Option<&UIImage>, style: UIBarButtonItemStyle, target: Option<&AnyObject>, action: Option<Sel>, ) -> Retained<Self>

Safety

target should be of the correct type.

action must be a valid selector.

pub unsafe fn initWithTitle_style_target_action( this: Allocated<Self>, title: Option<&NSString>, style: UIBarButtonItemStyle, target: Option<&AnyObject>, action: Option<Sel>, ) -> Retained<Self>

Safety

target should be of the correct type.

action must be a valid selector.

pub unsafe fn initWithBarButtonSystemItem_target_action( this: Allocated<Self>, system_item: UIBarButtonSystemItem, target: Option<&AnyObject>, action: Option<Sel>, ) -> Retained<Self>

Safety

target should be of the correct type.

action must be a valid selector.

pub fn initWithCustomView( this: Allocated<Self>, custom_view: &UIView, ) -> Retained<Self>

pub fn initWithBarButtonSystemItem_primaryAction( this: Allocated<Self>, system_item: UIBarButtonSystemItem, primary_action: Option<&UIAction>, ) -> Retained<Self>

Creates a bar button item for the given systemItem. The primaryAction is copied, and its title & image are ignored.

pub fn initWithPrimaryAction( this: Allocated<Self>, primary_action: Option<&UIAction>, ) -> Retained<Self>

Creates a plain-style bar button item from the properties of primaryAction. primaryAction is copied.

pub fn initWithBarButtonSystemItem_menu( this: Allocated<Self>, system_item: UIBarButtonSystemItem, menu: Option<&UIMenu>, ) -> Retained<Self>

Creates a bar button item for the given systemItem. The constructed item will present the menu immediately when touched.

pub fn initWithTitle_menu( this: Allocated<Self>, title: Option<&NSString>, menu: Option<&UIMenu>, ) -> Retained<Self>

Creates a plain-style bar button item with the given title. The constructed item will present the menu immediately when touched.

pub fn initWithImage_menu( this: Allocated<Self>, image: Option<&UIImage>, menu: Option<&UIMenu>, ) -> Retained<Self>

Creates a plain-style bar button item with the given image. The constructed item will present the menu immediately when touched.

pub fn initWithPrimaryAction_menu( this: Allocated<Self>, primary_action: Option<&UIAction>, menu: Option<&UIMenu>, ) -> Retained<Self>

Creates a plain-style bar button item from the properties of primaryAction. primaryAction is copied.

pub fn initWithBarButtonSystemItem_primaryAction_menu( this: Allocated<Self>, system_item: UIBarButtonSystemItem, primary_action: Option<&UIAction>, menu: Option<&UIMenu>, ) -> Retained<Self>

Creates a bar button item for the given systemItem. The primaryAction is copied, and its title & image are ignored.

pub unsafe fn initWithTitle_image_target_action_menu( this: Allocated<Self>, title: Option<&NSString>, image: Option<&UIImage>, target: Option<&AnyObject>, action: Option<Sel>, menu: Option<&UIMenu>, ) -> Retained<Self>

Creates a plain-style bar button item with the given title and image.

Safety

target should be of the correct type.

action must be a valid selector.

pub fn fixedSpaceItemOfWidth( width: CGFloat, mtm: MainThreadMarker, ) -> Retained<Self>

Construct a new fixed space item with the given width.

pub fn flexibleSpaceItem(mtm: MainThreadMarker) -> Retained<Self>

Construct a new flexible space item.

pub fn fixedSpaceItem(mtm: MainThreadMarker) -> Retained<Self>

Creates a new fixed space item of zero width.

A fixed space of 0 width separates the shared background used in navigation bars and toolbars to visually group items.

pub fn style(&self) -> UIBarButtonItemStyle

pub fn setStyle(&self, style: UIBarButtonItemStyle)

Setter for style.

pub fn width(&self) -> CGFloat

pub fn setWidth(&self, width: CGFloat)

Setter for width.

pub fn possibleTitles(&self) -> Option<Retained<NSSet<NSString>>>

pub fn setPossibleTitles(&self, possible_titles: Option<&NSSet<NSString>>)

Setter for possibleTitles.

This is copied when set.

pub fn customView(&self) -> Option<Retained<UIView>>

pub fn setCustomView(&self, custom_view: Option<&UIView>)

Setter for customView.

pub unsafe fn action(&self) -> Option<Sel>

Safety

You must ensure this is still alive.

pub unsafe fn setAction(&self, action: Option<Sel>)

Setter for action.

Safety

action must be a valid selector.

This is unretained, you must ensure the object is kept alive while in use.

pub fn target(&self) -> Option<Retained<AnyObject>>

pub unsafe fn setTarget(&self, target: Option<&AnyObject>)

Setter for target.

This is a [weak property][objc2::topics::weak_property].

Safety

target should be of the correct type.

pub fn primaryAction(&self) -> Option<Retained<UIAction>>

Set the primaryAction on this item, updating the title & image of the item if appropriate (primaryAction’s title is non-nil for the title update, primaryAction’s image is non-nil for the image update, and if this is not a system item). When primaryAction is non-nil, the target & action properties are ignored. If primaryAction is set to nil, the title & image properties are left unchanged.

pub fn setPrimaryAction(&self, primary_action: Option<&UIAction>)

Setter for primaryAction.

This is copied when set.

pub fn menu(&self) -> Option<Retained<UIMenu>>

When non-nil the menu is presented, the gesture used to trigger the menu is based on if the bar button item would normally trigger an action when tapped.

pub fn setMenu(&self, menu: Option<&UIMenu>)

Setter for menu.

This is copied when set.

pub fn preferredMenuElementOrder( &self, ) -> UIContextMenuConfigurationElementOrder

Preferred menu element ordering strategy for menus displayed by this button.

pub fn setPreferredMenuElementOrder( &self, preferred_menu_element_order: UIContextMenuConfigurationElementOrder, )

Setter for preferredMenuElementOrder.

pub fn changesSelectionAsPrimaryAction(&self) -> bool

Indicates if the button changes selection as its primary action. This shows the menu as options for selection if a menu is populated and no action when tapped is enabled. If no menu is provided and no action is enabled when tapped, the item is toggled on and off for the primary action.

pub fn setChangesSelectionAsPrimaryAction( &self, changes_selection_as_primary_action: bool, )

Setter for changesSelectionAsPrimaryAction.

pub fn isSelected(&self) -> bool

pub fn setSelected(&self, selected: bool)

Setter for isSelected.

pub fn isHidden(&self) -> bool

If the item should be hidden from display.

pub fn setHidden(&self, hidden: bool)

Setter for isHidden.

pub fn isSymbolAnimationEnabled(&self) -> bool

Whether or not symbol animations are enabled for this bar button item.

pub fn setSymbolAnimationEnabled(&self, symbol_animation_enabled: bool)

Setter for isSymbolAnimationEnabled.

pub fn menuRepresentation(&self) -> Option<Retained<UIMenuElement>>

A UIMenuElement that should substitute for the UIBarButtonItem when displayed in a menu.

pub fn setMenuRepresentation(&self, menu_representation: Option<&UIMenuElement>)

Setter for menuRepresentation.

This is copied when set.

pub fn hidesSharedBackground(&self) -> bool

A boolean value indicating whether the background this item may share with other items in the bar should be hidden.

Set this property to YES to prevent the standard shared background (typically using the Glass effect) from being drawn behind this bar button item.

This item will not be visually grouped with any other items, without the standard shared background. This property is ignored if the item is in a UIBarButtonItemGroup with more than one item. The default value is NO.

pub fn setHidesSharedBackground(&self, hides_shared_background: bool)

Setter for hidesSharedBackground.

pub fn sharesBackground(&self) -> bool

A boolean value indicating whether this bar button item can share a background with other items in a navigation bar or a toolbar.

When NO, This item will not be visually grouped with any other items.

This property is ignored if the item is in a UIBarButtonItemGroup with more than one item. The default value is YES.

pub fn setSharesBackground(&self, shares_background: bool)

Setter for sharesBackground.

pub fn identifier(&self) -> Option<Retained<NSString>>

An identifier used to match bar button items across transitions in a navigation bar or toolbar.

When the set of bar button items in a navigation bar or toolbar changes (for example, when pushing or popping view controllers), UIKit automatically animates the transition between the different sets of items. By default, UIKit uses heuristics based on item position and content to determine which items should be matched for these transitions.

Set this property with the same value on two different bar button items in different navigation item configurations to indicate that they should be treated as the same item during transitions. This allows for more natural animations when the visuals or function of an item changes across contexts.

The default value is nil, which means UIKit will use its default heuristics for transitions.

pub fn setIdentifier(&self, identifier: Option<&NSString>)

Setter for identifier.

This is copied when set.

pub fn creatingFixedGroup(&self) -> Retained<UIBarButtonItemGroup>

Create a fixed group containing this bar button item. UIBarButtonItems may only be in a single UIBarButtonItemGroup at a time, adding a bar button item to a group removes it from any previous group.

pub fn creatingMovableGroupWithCustomizationIdentifier( &self, customization_identifier: &NSString, ) -> Retained<UIBarButtonItemGroup>

Create a movable group containing this bar button item. UIBarButtonItems may only be in a single UIBarButtonItemGroup at a time, adding a bar button item to a group removes it from any previous group.

pub fn creatingOptionalGroupWithCustomizationIdentifier_inDefaultCustomization( &self, customization_identifier: &NSString, in_default_customization: bool, ) -> Retained<UIBarButtonItemGroup>

Create an optional group containing this bar button item. UIBarButtonItems may only be in a single UIBarButtonItemGroup at a time, adding a bar button item to a group removes it from any previous group.

pub fn setBackgroundImage_forState_barMetrics( &self, background_image: Option<&UIImage>, state: UIControlState, bar_metrics: UIBarMetrics, )

pub fn backgroundImageForState_barMetrics( &self, state: UIControlState, bar_metrics: UIBarMetrics, ) -> Option<Retained<UIImage>>

pub fn setBackgroundImage_forState_style_barMetrics( &self, background_image: Option<&UIImage>, state: UIControlState, style: UIBarButtonItemStyle, bar_metrics: UIBarMetrics, )

pub fn backgroundImageForState_style_barMetrics( &self, state: UIControlState, style: UIBarButtonItemStyle, bar_metrics: UIBarMetrics, ) -> Option<Retained<UIImage>>

pub fn tintColor(&self) -> Option<Retained<UIColor>>

pub fn setTintColor(&self, tint_color: Option<&UIColor>)

Setter for tintColor.

pub fn setBackgroundVerticalPositionAdjustment_forBarMetrics( &self, adjustment: CGFloat, bar_metrics: UIBarMetrics, )

pub fn backgroundVerticalPositionAdjustmentForBarMetrics( &self, bar_metrics: UIBarMetrics, ) -> CGFloat

pub fn setTitlePositionAdjustment_forBarMetrics( &self, adjustment: UIOffset, bar_metrics: UIBarMetrics, )

pub fn titlePositionAdjustmentForBarMetrics( &self, bar_metrics: UIBarMetrics, ) -> UIOffset

pub fn setBackButtonBackgroundImage_forState_barMetrics( &self, background_image: Option<&UIImage>, state: UIControlState, bar_metrics: UIBarMetrics, )

pub fn backButtonBackgroundImageForState_barMetrics( &self, state: UIControlState, bar_metrics: UIBarMetrics, ) -> Option<Retained<UIImage>>

pub fn setBackButtonTitlePositionAdjustment_forBarMetrics( &self, adjustment: UIOffset, bar_metrics: UIBarMetrics, )

pub fn backButtonTitlePositionAdjustmentForBarMetrics( &self, bar_metrics: UIBarMetrics, ) -> UIOffset

pub fn setBackButtonBackgroundVerticalPositionAdjustment_forBarMetrics( &self, adjustment: CGFloat, bar_metrics: UIBarMetrics, )

pub fn backButtonBackgroundVerticalPositionAdjustmentForBarMetrics( &self, bar_metrics: UIBarMetrics, ) -> CGFloat

impl UIBarButtonItem

Methods declared on superclass NSObject.

pub fn new(mtm: MainThreadMarker) -> Retained<Self>

impl UIBarButtonItem

This impl block contains no items.

SpringLoading.

impl UIBarButtonItem

pub fn removeAllSymbolEffects(&self)

Removes all symbol effects from the bar button item with default options and animation.

impl UIBarButtonItem

Badging.

pub fn badge(&self) -> Option<Retained<UIBarButtonItemBadge>>

Sets a badge on the bar button item. Currently only supported in navigation bars.

pub fn setBadge(&self, badge: Option<&UIBarButtonItemBadge>)

Setter for badge.

impl UIBarButtonItem

UIBarButtonItemGroup.

pub fn buttonGroup(&self) -> Option<Retained<UIBarButtonItemGroup>>

The group that the UIBarButtonItem is currently associated with, either as a member of the barButtonItems array or as that group’s representativeItem.

impl UIBarButtonItem

This impl block contains no items.

UIPopoverPresentationControllerSourceItem.