objc2_ui_kit

Struct UIButtonConfiguration

pub struct UIButtonConfiguration { /* private fields */ }

Apple’s documentation

Implementations

impl UIButtonConfiguration

pub fn plainButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn tintedButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn grayButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn filledButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn borderlessButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn borderedButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn borderedTintedButtonConfiguration( mtm: MainThreadMarker, ) -> Retained<Self>

pub fn borderedProminentButtonConfiguration( mtm: MainThreadMarker, ) -> Retained<Self>

pub fn glassButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn prominentGlassButtonConfiguration( mtm: MainThreadMarker, ) -> Retained<Self>

pub fn clearGlassButtonConfiguration(mtm: MainThreadMarker) -> Retained<Self>

pub fn prominentClearGlassButtonConfiguration( mtm: MainThreadMarker, ) -> Retained<Self>

pub unsafe fn new(mtm: MainThreadMarker) -> Retained<Self>

pub unsafe fn init(this: Allocated<Self>) -> Retained<Self>

pub fn updatedConfigurationForButton(&self, button: &UIButton) -> Retained<Self>

Returns a copy of the configuration updated based on the given button, by applying the configuration’s default values for that button’s state to any properties that have not been customized.

pub fn background(&self) -> Retained<UIBackgroundConfiguration>

A UIBackgroundConfiguration describing the button’s background. UIKit provides a value by default with values appropriate for a UIButton.

pub fn setBackground(&self, background: &UIBackgroundConfiguration)

Setter for background.

pub fn cornerStyle(&self) -> UIButtonConfigurationCornerStyle

The corner style controls how background.cornerRadius is interpreted by the button. Defaults to Dynamic.

pub fn setCornerStyle(&self, corner_style: UIButtonConfigurationCornerStyle)

Setter for cornerStyle.

pub fn buttonSize(&self) -> UIButtonConfigurationSize

Determines the metrics and ideal size of the button. Clients may resize the button arbitrarily regardless of this value.

pub fn setButtonSize(&self, button_size: UIButtonConfigurationSize)

Setter for buttonSize.

pub fn macIdiomStyle(&self) -> UIButtonConfigurationMacIdiomStyle

pub fn setMacIdiomStyle( &self, mac_idiom_style: UIButtonConfigurationMacIdiomStyle, )

Setter for macIdiomStyle.

pub fn baseForegroundColor(&self) -> Option<Retained<UIColor>>

The base color to use for foreground elements. This color may be modified before being passed to a transformer, and finally applied to specific elements. Setting nil will cede full control to the configuration to select a color appropriate to the style.

pub fn setBaseForegroundColor(&self, base_foreground_color: Option<&UIColor>)

Setter for baseForegroundColor.

pub fn baseBackgroundColor(&self) -> Option<Retained<UIColor>>

The base color to use for background elements. This color may be modified before being passed to a transformer, and finally applied to specific elements. Setting nil will cede full control to the configuration to select a color appropriate to the style.

pub fn setBaseBackgroundColor(&self, base_background_color: Option<&UIColor>)

Setter for baseBackgroundColor.

pub fn image(&self) -> Option<Retained<UIImage>>

pub fn setImage(&self, image: Option<&UIImage>)

Setter for image.

pub unsafe fn imageColorTransformer(&self) -> UIConfigurationColorTransformer

Safety

The returned block’s argument must be a valid pointer.

pub unsafe fn setImageColorTransformer( &self, image_color_transformer: UIConfigurationColorTransformer, )

Setter for imageColorTransformer.

This is copied when set.

Safety

image_color_transformer must be a valid pointer or null.

pub fn preferredSymbolConfigurationForImage( &self, ) -> Option<Retained<UIImageSymbolConfiguration>>

pub fn setPreferredSymbolConfigurationForImage( &self, preferred_symbol_configuration_for_image: Option<&UIImageSymbolConfiguration>, )

Setter for preferredSymbolConfigurationForImage.

This is copied when set.

pub fn symbolContentTransition( &self, ) -> Option<Retained<UISymbolContentTransition>>

The symbol content transition to use when transitioning across symbol images. Defaults to nil, meaning no symbol content transition should occur.

pub fn setSymbolContentTransition( &self, symbol_content_transition: Option<&UISymbolContentTransition>, )

Setter for symbolContentTransition.

pub fn showsActivityIndicator(&self) -> bool

Shows an activity indicator in place of an image. Its placement is controlled by the imagePlacement property.

pub fn setShowsActivityIndicator(&self, shows_activity_indicator: bool)

Setter for showsActivityIndicator.

pub unsafe fn activityIndicatorColorTransformer( &self, ) -> UIConfigurationColorTransformer

Safety

The returned block’s argument must be a valid pointer.

pub unsafe fn setActivityIndicatorColorTransformer( &self, activity_indicator_color_transformer: UIConfigurationColorTransformer, )

Setter for activityIndicatorColorTransformer.

This is copied when set.

Safety

activity_indicator_color_transformer must be a valid pointer or null.

pub fn title(&self) -> Option<Retained<NSString>>

pub fn setTitle(&self, title: Option<&NSString>)

Setter for title.

This is copied when set.

pub fn attributedTitle(&self) -> Option<Retained<NSAttributedString>>

pub fn setAttributedTitle(&self, attributed_title: Option<&NSAttributedString>)

Setter for attributedTitle.

This is copied when set.

pub unsafe fn titleTextAttributesTransformer( &self, ) -> UIConfigurationTextAttributesTransformer

Safety

The returned block’s argument must be a valid pointer.

pub unsafe fn setTitleTextAttributesTransformer( &self, title_text_attributes_transformer: UIConfigurationTextAttributesTransformer, )

Setter for titleTextAttributesTransformer.

This is copied when set.

Safety

title_text_attributes_transformer must be a valid pointer or null.

pub fn titleLineBreakMode(&self) -> NSLineBreakMode

Default is WordWrapping. WordWrapping and CharWrapping both allow for multi-line text, other modes will restrict the title to a single line of text.

pub fn setTitleLineBreakMode(&self, title_line_break_mode: NSLineBreakMode)

Setter for titleLineBreakMode.

pub fn subtitle(&self) -> Option<Retained<NSString>>

pub fn setSubtitle(&self, subtitle: Option<&NSString>)

Setter for subtitle.

This is copied when set.

pub fn attributedSubtitle(&self) -> Option<Retained<NSAttributedString>>

pub fn setAttributedSubtitle( &self, attributed_subtitle: Option<&NSAttributedString>, )

Setter for attributedSubtitle.

This is copied when set.

pub unsafe fn subtitleTextAttributesTransformer( &self, ) -> UIConfigurationTextAttributesTransformer

Safety

The returned block’s argument must be a valid pointer.

pub unsafe fn setSubtitleTextAttributesTransformer( &self, subtitle_text_attributes_transformer: UIConfigurationTextAttributesTransformer, )

Setter for subtitleTextAttributesTransformer.

This is copied when set.

Safety

subtitle_text_attributes_transformer must be a valid pointer or null.

pub fn subtitleLineBreakMode(&self) -> NSLineBreakMode

Default is WordWrapping. WordWrapping and CharWrapping both allow for multi-line text, other modes will restrict the subtitle to a single line of text.

pub fn setSubtitleLineBreakMode( &self, subtitle_line_break_mode: NSLineBreakMode, )

Setter for subtitleLineBreakMode.

pub fn indicator(&self) -> UIButtonConfigurationIndicator

What kind of indicator should the button show. Default value is .automatic.

pub fn setIndicator(&self, indicator: UIButtonConfigurationIndicator)

Setter for indicator.

pub unsafe fn indicatorColorTransformer( &self, ) -> UIConfigurationColorTransformer

Safety

The returned block’s argument must be a valid pointer.

pub unsafe fn setIndicatorColorTransformer( &self, indicator_color_transformer: UIConfigurationColorTransformer, )

Setter for indicatorColorTransformer.

This is copied when set.

Safety

indicator_color_transformer must be a valid pointer or null.

pub fn contentInsets(&self) -> NSDirectionalEdgeInsets

Insets from the bounds of the button to create the content region. Defaults styles provide insets based on the button size.

pub fn setContentInsets(&self, content_insets: NSDirectionalEdgeInsets)

Setter for contentInsets.

pub fn setDefaultContentInsets(&self)

Restore the default content insets.

pub fn imagePlacement(&self) -> NSDirectionalRectEdge

Defaults to Leading, only single edge values (top/leading/bottom/trailing) are supported.

pub fn setImagePlacement(&self, image_placement: NSDirectionalRectEdge)

Setter for imagePlacement.

pub fn imagePadding(&self) -> CGFloat

When a button has both image and text content, this value is the padding between the image and the text.

pub fn setImagePadding(&self, image_padding: CGFloat)

Setter for imagePadding.

pub fn titlePadding(&self) -> CGFloat

When a button has both a title & subtitle, this value is the padding between those titles.

pub fn setTitlePadding(&self, title_padding: CGFloat)

Setter for titlePadding.

pub fn titleAlignment(&self) -> UIButtonConfigurationTitleAlignment

The alignment to use for relative layout between title & subtitle.

pub fn setTitleAlignment( &self, title_alignment: UIButtonConfigurationTitleAlignment, )

Setter for titleAlignment.

pub fn automaticallyUpdateForSelection(&self) -> bool

If the style should automatically update when the button is selected. Default varies by style. Disable to customize selection behavior.

pub fn setAutomaticallyUpdateForSelection( &self, automatically_update_for_selection: bool, )

Setter for automaticallyUpdateForSelection.