objc2_ui_kit

Struct UIDocumentViewController

pub struct UIDocumentViewController { /* private fields */ }

Apple’s documentation

Implementations

impl UIDocumentViewController

pub fn initWithDocument( this: Allocated<Self>, document: Option<&UIDocument>, ) -> Retained<Self>

pub fn document(&self) -> Option<Retained<UIDocument>>

pub fn setDocument(&self, document: Option<&UIDocument>)

Setter for document.

pub fn launchOptions(&self) -> Retained<UIDocumentViewControllerLaunchOptions>

Properties to configure the view controller when no document is open.

pub fn setLaunchOptions( &self, launch_options: &UIDocumentViewControllerLaunchOptions, )

Setter for launchOptions.

pub fn navigationItemDidUpdate(&self)

When this view controller updates its navigation item, this method will be called, allowing subclasses to apply any kind of customization you might want.

pub fn openDocumentWithCompletionHandler( &self, completion_handler: &DynBlock<dyn Fn(Bool)>, )

Opens the current document. This method is optional. If the document is not opened by the time the view controller becomes visible, the view controller will take care of opening the document. If the document is already opened, the completion handler will be called as if opening the document succeeded.

pub fn documentDidOpen(&self)

Notifies subclasses that the specified document was opened. This method will be called by the system after opening the document initially or after the presented document was changed.

pub fn undoRedoItemGroup(&self) -> Retained<UIBarButtonItemGroup>

MARK: Custom Navigation Items

impl UIDocumentViewController

Methods declared on superclass UIViewController.

pub fn initWithNibName_bundle( this: Allocated<Self>, nib_name_or_nil: Option<&NSString>, nib_bundle_or_nil: Option<&NSBundle>, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

impl UIDocumentViewController

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new(mtm: MainThreadMarker) -> Retained<Self>