objc2_ui_kit

Struct UIImage

pub struct UIImage { /* private fields */ }

Apple’s documentation

Implementations

impl UIImage

This impl block contains no items.

UIAccessibility.

impl UIImage

pub fn systemImageNamed(name: &NSString) -> Option<Retained<UIImage>>

pub fn systemImageNamed_withConfiguration( name: &NSString, configuration: Option<&UIImageConfiguration>, ) -> Option<Retained<UIImage>>

pub fn systemImageNamed_compatibleWithTraitCollection( name: &NSString, trait_collection: Option<&UITraitCollection>, ) -> Option<Retained<UIImage>>

pub fn systemImageNamed_variableValue_withConfiguration( name: &NSString, value: c_double, configuration: Option<&UIImageConfiguration>, ) -> Option<Retained<UIImage>>

Retrieve a system-provided image with the specified name and variable value (between 0 and 1).

This will only return system-provided images. If you want a custom image as defined in your own catalogs, you should use +imageNamed:inBundle:variableValue:withConfiguration:. Returns nilif an image with specified name doesn’t exist.

pub fn imageNamed(name: &NSString) -> Option<Retained<UIImage>>

pub fn imageNamed_inBundle_withConfiguration( name: &NSString, bundle: Option<&NSBundle>, configuration: Option<&UIImageConfiguration>, ) -> Option<Retained<UIImage>>

pub fn imageNamed_inBundle_compatibleWithTraitCollection( name: &NSString, bundle: Option<&NSBundle>, trait_collection: Option<&UITraitCollection>, ) -> Option<Retained<UIImage>>

pub fn imageNamed_inBundle_variableValue_withConfiguration( name: &NSString, bundle: Option<&NSBundle>, value: c_double, configuration: Option<&UIImageConfiguration>, ) -> Option<Retained<UIImage>>

Retrieve a image with the specified name and variable value (between 0 and 1).

This will only return custom images defined in your own catalogs. If you want a system-provided image, you should use +systemImageNamed:variableValue:withConfiguration:. Returns nilif an image with specified name doesn’t exist.

pub fn imageWithContentsOfFile(path: &NSString) -> Option<Retained<UIImage>>

pub fn imageWithData(data: &NSData) -> Option<Retained<UIImage>>

pub fn imageWithData_scale( data: &NSData, scale: CGFloat, ) -> Option<Retained<UIImage>>

pub fn imageWithCGImage(cg_image: &CGImage) -> Retained<UIImage>

pub fn imageWithCGImage_scale_orientation( cg_image: &CGImage, scale: CGFloat, orientation: UIImageOrientation, ) -> Retained<UIImage>

pub fn imageWithCIImage(ci_image: &CIImage) -> Retained<UIImage>

pub fn imageWithCIImage_scale_orientation( ci_image: &CIImage, scale: CGFloat, orientation: UIImageOrientation, ) -> Retained<UIImage>

pub fn initWithContentsOfFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<Self>>

pub fn initWithData( this: Allocated<Self>, data: &NSData, ) -> Option<Retained<Self>>

pub fn initWithData_scale( this: Allocated<Self>, data: &NSData, scale: CGFloat, ) -> Option<Retained<Self>>

pub fn initWithCGImage( this: Allocated<Self>, cg_image: &CGImage, ) -> Retained<Self>

pub fn initWithCGImage_scale_orientation( this: Allocated<Self>, cg_image: &CGImage, scale: CGFloat, orientation: UIImageOrientation, ) -> Retained<Self>

pub fn initWithCIImage( this: Allocated<Self>, ci_image: &CIImage, ) -> Retained<Self>

pub fn initWithCIImage_scale_orientation( this: Allocated<Self>, ci_image: &CIImage, scale: CGFloat, orientation: UIImageOrientation, ) -> Retained<Self>

pub unsafe fn size(&self) -> CGSize

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn CGImage(&self) -> Option<Retained<CGImage>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn CIImage(&self) -> Option<Retained<CIImage>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn imageOrientation(&self) -> UIImageOrientation

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn scale(&self) -> CGFloat

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn isSymbolImage(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub fn animatedImageNamed_duration( name: &NSString, duration: NSTimeInterval, ) -> Option<Retained<UIImage>>

pub fn animatedResizableImageNamed_capInsets_duration( name: &NSString, cap_insets: UIEdgeInsets, duration: NSTimeInterval, ) -> Option<Retained<UIImage>>

pub fn animatedResizableImageNamed_capInsets_resizingMode_duration( name: &NSString, cap_insets: UIEdgeInsets, resizing_mode: UIImageResizingMode, duration: NSTimeInterval, ) -> Option<Retained<UIImage>>

pub fn animatedImageWithImages_duration( images: &NSArray<UIImage>, duration: NSTimeInterval, ) -> Option<Retained<UIImage>>

pub unsafe fn images(&self) -> Option<Retained<NSArray<UIImage>>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn duration(&self) -> NSTimeInterval

This property is not atomic.

Safety

This might not be thread-safe.

pub fn drawAtPoint(&self, point: CGPoint)

pub fn drawAtPoint_blendMode_alpha( &self, point: CGPoint, blend_mode: CGBlendMode, alpha: CGFloat, )

pub fn drawInRect(&self, rect: CGRect)

pub fn drawInRect_blendMode_alpha( &self, rect: CGRect, blend_mode: CGBlendMode, alpha: CGFloat, )

pub fn drawAsPatternInRect(&self, rect: CGRect)

pub fn resizableImageWithCapInsets( &self, cap_insets: UIEdgeInsets, ) -> Retained<UIImage>

pub fn resizableImageWithCapInsets_resizingMode( &self, cap_insets: UIEdgeInsets, resizing_mode: UIImageResizingMode, ) -> Retained<UIImage>

pub unsafe fn capInsets(&self) -> UIEdgeInsets

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn resizingMode(&self) -> UIImageResizingMode

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageWithAlignmentRectInsets( &self, alignment_insets: UIEdgeInsets, ) -> Retained<UIImage>

pub unsafe fn alignmentRectInsets(&self) -> UIEdgeInsets

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageWithRenderingMode( &self, rendering_mode: UIImageRenderingMode, ) -> Retained<UIImage>

pub unsafe fn renderingMode(&self) -> UIImageRenderingMode

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn imageRendererFormat( &self, ) -> Retained<UIGraphicsImageRendererFormat>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn traitCollection(&self) -> Retained<UITraitCollection>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn imageAsset(&self) -> Option<Retained<UIImageAsset>>

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageFlippedForRightToLeftLayoutDirection(&self) -> Retained<UIImage>

pub unsafe fn flipsForRightToLeftLayoutDirection(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageWithHorizontallyFlippedOrientation(&self) -> Retained<UIImage>

pub unsafe fn baselineOffsetFromBottom(&self) -> CGFloat

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn hasBaseline(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageWithBaselineOffsetFromBottom( &self, baseline_offset: CGFloat, ) -> Retained<UIImage>

pub fn imageWithoutBaseline(&self) -> Retained<UIImage>

pub unsafe fn configuration(&self) -> Option<Retained<UIImageConfiguration>>

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageWithConfiguration( &self, configuration: &UIImageConfiguration, ) -> Retained<UIImage>

pub unsafe fn symbolConfiguration( &self, ) -> Option<Retained<UIImageSymbolConfiguration>>

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageByApplyingSymbolConfiguration( &self, configuration: &UIImageSymbolConfiguration, ) -> Option<Retained<UIImage>>

pub fn imageWithTintColor(&self, color: &UIColor) -> Retained<UIImage>

pub fn imageWithTintColor_renderingMode( &self, color: &UIColor, rendering_mode: UIImageRenderingMode, ) -> Retained<UIImage>

pub fn imageByPreparingForDisplay(&self) -> Option<Retained<UIImage>>

Decodes an image synchronously and provides a new one for display in views and animations.

Returns: A new version of the image object for display. If the system can’t decode the image, this method returns nil.

Note: The prepared UIImage is not related to the original image. If the properties of the screen (such as its resolution or color gamut) change, or if the image is displayed on a different screen that the one it was prepared for, it may not render correctly.

pub fn prepareForDisplayWithCompletionHandler( &self, completion_handler: &DynBlock<dyn Fn(*mut UIImage)>, )

Decodes an image asynchronously and provides a new one for display in views and animations.

The completion handler will be invoked on a private queue. Be sure to return to the main queue before assigning the prepared image to an image view.

Parameter completionHandler: A block to invoke with the prepared image. If preparation failed (for example, because the image data is corrupt), imagewill be nil.

Note: The prepared UIImage is not related to the original image. If the properties of the screen (such as its resolution or color gamut) change, or if the image is displayed on a different screen that the one it was prepared for, it may not render correctly.

pub fn imageByPreparingThumbnailOfSize( &self, size: CGSize, ) -> Option<Retained<UIImage>>

pub fn prepareThumbnailOfSize_completionHandler( &self, size: CGSize, completion_handler: &DynBlock<dyn Fn(*mut UIImage)>, )

pub unsafe fn isHighDynamicRange(&self) -> bool

Indicates that this image is tagged for display of high dynamic range content.

This property is not atomic.

Safety

This might not be thread-safe.

pub fn imageRestrictedToStandardDynamicRange(&self) -> Retained<UIImage>

Returns a new image that will render within the standard range.

impl UIImage

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new() -> Retained<Self>

impl UIImage

PreconfiguredSystemImages.

pub fn actionsImage() -> Retained<UIImage>

pub fn addImage() -> Retained<UIImage>

pub fn removeImage() -> Retained<UIImage>

pub fn checkmarkImage() -> Retained<UIImage>

pub fn strokedCheckmarkImage() -> Retained<UIImage>

impl UIImage

This impl block contains no items.

NSItemProvider.

impl UIImage

UIImageDeprecated.

pub fn stretchableImageWithLeftCapWidth_topCapHeight( &self, left_cap_width: NSInteger, top_cap_height: NSInteger, ) -> Retained<UIImage>

pub unsafe fn leftCapWidth(&self) -> NSInteger

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn topCapHeight(&self) -> NSInteger

This property is not atomic.

Safety

This might not be thread-safe.

impl UIImage

pub fn png_representation(&self) -> Option<Retained<NSData>>

return image as PNG. May return nil if image has no CGImageRef or invalid bitmap format

pub fn jpeg_representation( &self, compression_quality: CGFloat, ) -> Option<Retained<NSData>>

return image as JPEG. May return nil if image has no CGImageRef or invalid bitmap format. compression is 0(most)..1(least)

pub fn heic_representation(&self) -> Option<Retained<NSData>>

Returns HEIC data representing the image, or nil if such a representation could not be generated. HEIC is recommended for efficiently storing all kinds of images, including those with high dynamic range content.

impl UIImage

pub unsafe fn write_to_saved_photos_album( &self, completion_target: Option<&AnyObject>, completion_selector: Option<Sel>, context_info: *mut c_void, )

Safety

completion_target should be of the correct type.

completion_selector must be a valid selector.

context_info must be a valid pointer or null.