objc2_ui_kit

Struct UIKeyboardHIDUsage

#[repr(transparent)]
pub struct UIKeyboardHIDUsage(pub CFIndex);

This file defines a table of standardized key codes, called “HID usages” used by USB keyboards to identify individual keys.

A general note on Usages and languages: Due to the variation of keyboards from language to language, it is not feasible to specify exact key mappings for every language. Where this list is not specific for a key function in a language, the closest equivalent key position should be used, so that a keyboard may be modified for a different language by simply printing different keycaps. One example is the Y key on a North American keyboard. In Germany this is typically Z. Rather than changing the keyboard firmware to put the Z Usage into that place in the descriptor list, the vendor uses the Y Usage on both the North American and German keyboards. This continues to be the existing practice in the industry, in order to minimize the number of changes to the electronics to accommodate other languages.

The following constants are from the USB ‘HID Usage Tables’ specification, revision 1.1rc3.

See also Apple’s documentation

Tuple Fields
0: CFIndex
Implementations

impl UIKeyboardHIDUsage

pub const KeyboardErrorRollOver: Self

pub const KeyboardPOSTFail: Self

pub const KeyboardErrorUndefined: Self

pub const KeyboardA: Self

pub const KeyboardB: Self

pub const KeyboardC: Self

pub const KeyboardD: Self

pub const KeyboardE: Self

pub const KeyboardF: Self

pub const KeyboardG: Self

pub const KeyboardH: Self

pub const KeyboardI: Self

pub const KeyboardJ: Self

pub const KeyboardK: Self

pub const KeyboardL: Self

pub const KeyboardM: Self

pub const KeyboardN: Self

pub const KeyboardO: Self

pub const KeyboardP: Self

pub const KeyboardQ: Self

pub const KeyboardR: Self

pub const KeyboardS: Self

pub const KeyboardT: Self

pub const KeyboardU: Self

pub const KeyboardV: Self

pub const KeyboardW: Self

pub const KeyboardX: Self

pub const KeyboardY: Self

pub const KeyboardZ: Self

pub const Keyboard1: Self

pub const Keyboard2: Self

pub const Keyboard3: Self

pub const Keyboard4: Self

pub const Keyboard5: Self

pub const Keyboard6: Self

pub const Keyboard7: Self

pub const Keyboard8: Self

pub const Keyboard9: Self

pub const Keyboard0: Self

pub const KeyboardReturnOrEnter: Self

pub const KeyboardEscape: Self

pub const KeyboardDeleteOrBackspace: Self

pub const KeyboardTab: Self

pub const KeyboardSpacebar: Self

pub const KeyboardHyphen: Self

pub const KeyboardEqualSign: Self

pub const KeyboardOpenBracket: Self

pub const KeyboardCloseBracket: Self

pub const KeyboardBackslash: Self

pub const KeyboardNonUSPound: Self

pub const KeyboardSemicolon: Self

pub const KeyboardQuote: Self

pub const KeyboardGraveAccentAndTilde: Self

pub const KeyboardComma: Self

pub const KeyboardPeriod: Self

pub const KeyboardSlash: Self

pub const KeyboardCapsLock: Self

pub const KeyboardF1: Self

pub const KeyboardF2: Self

pub const KeyboardF3: Self

pub const KeyboardF4: Self

pub const KeyboardF5: Self

pub const KeyboardF6: Self

pub const KeyboardF7: Self

pub const KeyboardF8: Self

pub const KeyboardF9: Self

pub const KeyboardF10: Self

pub const KeyboardF11: Self

pub const KeyboardF12: Self

pub const KeyboardPrintScreen: Self

pub const KeyboardScrollLock: Self

pub const KeyboardPause: Self

pub const KeyboardInsert: Self

pub const KeyboardHome: Self

pub const KeyboardPageUp: Self

pub const KeyboardDeleteForward: Self

pub const KeyboardEnd: Self

pub const KeyboardPageDown: Self

pub const KeyboardRightArrow: Self

pub const KeyboardLeftArrow: Self

pub const KeyboardDownArrow: Self

pub const KeyboardUpArrow: Self

pub const KeypadNumLock: Self

pub const KeypadSlash: Self

pub const KeypadAsterisk: Self

pub const KeypadHyphen: Self

pub const KeypadPlus: Self

pub const KeypadEnter: Self

pub const Keypad1: Self

pub const Keypad2: Self

pub const Keypad3: Self

pub const Keypad4: Self

pub const Keypad5: Self

pub const Keypad6: Self

pub const Keypad7: Self

pub const Keypad8: Self

pub const Keypad9: Self

pub const Keypad0: Self

pub const KeypadPeriod: Self

pub const KeyboardNonUSBackslash: Self

pub const KeyboardApplication: Self

pub const KeyboardPower: Self

pub const KeypadEqualSign: Self

pub const KeyboardF13: Self

pub const KeyboardF14: Self

pub const KeyboardF15: Self

pub const KeyboardF16: Self

pub const KeyboardF17: Self

pub const KeyboardF18: Self

pub const KeyboardF19: Self

pub const KeyboardF20: Self

pub const KeyboardF21: Self

pub const KeyboardF22: Self

pub const KeyboardF23: Self

pub const KeyboardF24: Self

pub const KeyboardExecute: Self

pub const KeyboardHelp: Self

pub const KeyboardMenu: Self

pub const KeyboardSelect: Self

pub const KeyboardStop: Self

pub const KeyboardAgain: Self

pub const KeyboardUndo: Self

pub const KeyboardCut: Self

pub const KeyboardCopy: Self

pub const KeyboardPaste: Self

pub const KeyboardFind: Self

pub const KeyboardMute: Self

pub const KeyboardVolumeUp: Self

pub const KeyboardVolumeDown: Self

pub const KeyboardLockingCapsLock: Self

pub const KeyboardLockingNumLock: Self

pub const KeyboardLockingScrollLock: Self

pub const KeypadComma: Self

pub const KeypadEqualSignAS400: Self

pub const KeyboardInternational1: Self

pub const KeyboardInternational2: Self

pub const KeyboardInternational3: Self

pub const KeyboardInternational4: Self

pub const KeyboardInternational5: Self

pub const KeyboardInternational6: Self

pub const KeyboardInternational7: Self

pub const KeyboardInternational8: Self

pub const KeyboardInternational9: Self

pub const KeyboardLANG1: Self

pub const KeyboardLANG2: Self

pub const KeyboardLANG3: Self

pub const KeyboardLANG4: Self

pub const KeyboardLANG5: Self

pub const KeyboardLANG6: Self

pub const KeyboardLANG7: Self

pub const KeyboardLANG8: Self

pub const KeyboardLANG9: Self

pub const KeyboardAlternateErase: Self

pub const KeyboardSysReqOrAttention: Self

pub const KeyboardCancel: Self

pub const KeyboardClear: Self

pub const KeyboardPrior: Self

pub const KeyboardReturn: Self

pub const KeyboardSeparator: Self

pub const KeyboardOut: Self

pub const KeyboardOper: Self

pub const KeyboardClearOrAgain: Self

pub const KeyboardCrSelOrProps: Self

pub const KeyboardExSel: Self

pub const KeyboardLeftControl: Self

pub const KeyboardLeftShift: Self

pub const KeyboardLeftAlt: Self

pub const KeyboardLeftGUI: Self

pub const KeyboardRightControl: Self

pub const KeyboardRightShift: Self

pub const KeyboardRightAlt: Self

pub const KeyboardRightGUI: Self

pub const Keyboard_Reserved: Self

pub const KeyboardHangul: Self

pub const KeyboardHanja: Self

pub const KeyboardKanaSwitch: Self

pub const KeyboardAlphanumericSwitch: Self

pub const KeyboardKatakana: Self

pub const KeyboardHiragana: Self

pub const KeyboardZenkakuHankakuKanji: Self