objc2_ui_kit

Struct UINavigationItem

pub struct UINavigationItem { /* private fields */ }

Apple’s documentation

Implementations

impl UINavigationItem

pub fn initWithTitle(this: Allocated<Self>, title: &NSString) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

pub fn title(&self) -> Option<Retained<NSString>>

Title when topmost on the stack. default is nil

pub fn setTitle(&self, title: Option<&NSString>)

Setter for title.

This is copied when set.

pub fn attributedTitle(&self) -> Option<Retained<NSAttributedString>>

An attributed string that is rendered as the title in the navigation bar.

If titleView is non-nil, this property is ignored.

pub fn setAttributedTitle(&self, attributed_title: Option<&NSAttributedString>)

Setter for attributedTitle.

This is copied when set.

pub fn titleView(&self) -> Option<Retained<UIView>>

Custom view to use in lieu of a title. May be sized horizontally. Only used when item is topmost on the stack.

pub fn setTitleView(&self, title_view: Option<&UIView>)

Setter for titleView.

pub fn subtitle(&self) -> Option<Retained<NSString>>

A string to display as the subtitle in the navigation bar.

If attributedSubtitle is non-nil, this property just returns the String representation of the attributedString. If subtitleView is non-nil, this property is ignored.

pub fn setSubtitle(&self, subtitle: Option<&NSString>)

Setter for subtitle.

This is copied when set.

pub fn attributedSubtitle(&self) -> Option<Retained<NSAttributedString>>

An attributed string to display as the subtitle in the navigation bar.

If non-nil, this property takes precedence over the subtitle property. If subtitleView is non-nil, this property is ignored. If titleView is non-nil, this property is ignored.

pub fn setAttributedSubtitle( &self, attributed_subtitle: Option<&NSAttributedString>, )

Setter for attributedSubtitle.

This is copied when set.

pub fn subtitleView(&self) -> Option<Retained<UIView>>

A custom view to display below the title in the navigation bar.

If non-nil, this property takes precedence over the subtitle and attributedSubtitle properties. The view’s layout constraints will determine its size, or the view may override sizeThatFits(_:) to return its desired size.

pub fn setSubtitleView(&self, subtitle_view: Option<&UIView>)

Setter for subtitleView.

This is copied when set.

pub fn largeTitle(&self) -> Option<Retained<NSString>>

String to be used as the large title.

When nil, the navigation bar will use the navigation item’s current title.

pub fn setLargeTitle(&self, large_title: Option<&NSString>)

Setter for largeTitle.

This is copied when set.

pub fn largeSubtitle(&self) -> Option<Retained<NSString>>

String to be rendered below the large title.

When nil, the navigation bar will fall back to the subtitle.

pub fn setLargeSubtitle(&self, large_subtitle: Option<&NSString>)

Setter for largeSubtitle.

This is copied when set.

pub fn largeAttributedSubtitle(&self) -> Option<Retained<NSAttributedString>>

An attributed string to be rendered below the large title.

When nil, the navigation bar will fall back to the largeSubtitle. If a largeSubtitleView is set, this property is ignored.

pub fn setLargeAttributedSubtitle( &self, large_attributed_subtitle: Option<&NSAttributedString>, )

Setter for largeAttributedSubtitle.

This is copied when set.

pub fn largeSubtitleView(&self) -> Option<Retained<UIView>>

A custom view to display below the large title.

When non-nil, this takes precedence over any other subtitle. The view’s layout constraints will determine its size, or the view may override sizeThatFits(_:) to return its desired size.

pub fn setLargeSubtitleView(&self, large_subtitle_view: Option<&UIView>)

Setter for largeSubtitleView.

pub fn prompt(&self) -> Option<Retained<NSString>>

Explanatory text to display above the navigation bar buttons.

pub fn setPrompt(&self, prompt: Option<&NSString>)

Setter for prompt.

This is copied when set.

pub fn backBarButtonItem(&self) -> Option<Retained<UIBarButtonItem>>

Bar button item to use for the back button when this item is the navigation bar’s backItem.

pub fn setBackBarButtonItem( &self, back_bar_button_item: Option<&UIBarButtonItem>, )

Setter for backBarButtonItem.

pub fn backButtonTitle(&self) -> Option<Retained<NSString>>

Title to use when this item is the navigation bar’s backItem. Default is nil. backBarButtonItem takes precedence if both are specified.

pub fn setBackButtonTitle(&self, back_button_title: Option<&NSString>)

Setter for backButtonTitle.

This is copied when set.

pub fn hidesBackButton(&self) -> bool

If YES, this navigation item will hide the back button when it’s on top of the stack.

pub fn setHidesBackButton(&self, hides_back_button: bool)

Setter for hidesBackButton.

pub fn setHidesBackButton_animated( &self, hides_back_button: bool, animated: bool, )

pub fn backButtonDisplayMode(&self) -> UINavigationItemBackButtonDisplayMode

Controls how the back button sources its title.

pub fn setBackButtonDisplayMode( &self, back_button_display_mode: UINavigationItemBackButtonDisplayMode, )

Setter for backButtonDisplayMode.

pub fn backAction(&self) -> Option<Retained<UIAction>>

Replaces the back action for the navigation bar. if a back button is naturally present, this replaces only its action, otherwise a back button will be synthesized with the image or title from the action.

pub fn setBackAction(&self, back_action: Option<&UIAction>)

Setter for backAction.

This is copied when set.

pub unsafe fn titleMenuProvider( &self, ) -> *mut DynBlock<dyn Fn(NonNull<NSArray<UIMenuElement>>) -> *mut UIMenu>

When non-nil, UIKit will generate suggestedActions and call this block to generate a menu that is displayed from the title.

Safety

The returned block’s argument must be a valid pointer.

pub unsafe fn setTitleMenuProvider( &self, title_menu_provider: Option<&DynBlock<dyn Fn(NonNull<NSArray<UIMenuElement>>) -> *mut UIMenu>>, )

Setter for titleMenuProvider.

This is copied when set.

Safety

title_menu_provider block’s return must be a valid pointer or null.

pub fn renameDelegate( &self, ) -> Option<Retained<ProtocolObject<dyn UINavigationItemRenameDelegate>>>

When non-nil, enables the UINavigationBar to respond to the rename: action by providing an inline UI. Setting this property to nil while a rename operation is in progress will cancel the operation immediately.

pub fn setRenameDelegate( &self, rename_delegate: Option<&ProtocolObject<dyn UINavigationItemRenameDelegate>>, )

Setter for renameDelegate.

This is a [weak property][objc2::topics::weak_property].

pub fn documentProperties(&self) -> Option<Retained<UIDocumentProperties>>

When non-nil, UIKit will use the metadata provided to generate additional controls displayed from the title.

pub fn setDocumentProperties( &self, document_properties: Option<&UIDocumentProperties>, )

Setter for documentProperties.

pub fn leftBarButtonItems(&self) -> Option<Retained<NSArray<UIBarButtonItem>>>

pub fn setLeftBarButtonItems( &self, left_bar_button_items: Option<&NSArray<UIBarButtonItem>>, )

Setter for leftBarButtonItems.

This is copied when set.

pub fn rightBarButtonItems(&self) -> Option<Retained<NSArray<UIBarButtonItem>>>

pub fn setRightBarButtonItems( &self, right_bar_button_items: Option<&NSArray<UIBarButtonItem>>, )

Setter for rightBarButtonItems.

This is copied when set.

pub fn setLeftBarButtonItems_animated( &self, items: Option<&NSArray<UIBarButtonItem>>, animated: bool, )

pub fn setRightBarButtonItems_animated( &self, items: Option<&NSArray<UIBarButtonItem>>, animated: bool, )

pub fn leftItemsSupplementBackButton(&self) -> bool

pub fn setLeftItemsSupplementBackButton( &self, left_items_supplement_back_button: bool, )

Setter for leftItemsSupplementBackButton.

pub fn leftBarButtonItem(&self) -> Option<Retained<UIBarButtonItem>>

Some navigation items want to display a custom left or right item when they’re on top of the stack. A custom left item replaces the regular back button unless you set leftItemsSupplementBackButton to YES

pub fn setLeftBarButtonItem( &self, left_bar_button_item: Option<&UIBarButtonItem>, )

Setter for leftBarButtonItem.

pub fn rightBarButtonItem(&self) -> Option<Retained<UIBarButtonItem>>

pub fn setRightBarButtonItem( &self, right_bar_button_item: Option<&UIBarButtonItem>, )

Setter for rightBarButtonItem.

pub fn setLeftBarButtonItem_animated( &self, item: Option<&UIBarButtonItem>, animated: bool, )

pub fn setRightBarButtonItem_animated( &self, item: Option<&UIBarButtonItem>, animated: bool, )

pub fn customizationIdentifier(&self) -> Option<Retained<NSString>>

Setting a non-nil customizationIdentifier enables customization and UIKit will automatically save & restore customizations based on this identifier. The identifier must be unique within the scope of the application.

pub fn setCustomizationIdentifier( &self, customization_identifier: Option<&NSString>, )

Setter for customizationIdentifier.

This is copied when set.

pub fn leadingItemGroups(&self) -> Retained<NSArray<UIBarButtonItemGroup>>

UIBarButtonItemGroups to be displayed in the leading section of the navigation bar. Items set via this method will replace items set via .leftBarButtonItem or .leftBarButtonItems.

pub fn setLeadingItemGroups( &self, leading_item_groups: &NSArray<UIBarButtonItemGroup>, )

Setter for leadingItemGroups.

This is copied when set.

pub fn centerItemGroups(&self) -> Retained<NSArray<UIBarButtonItemGroup>>

Customizable UIBarButtonItemGroups that will be placed in the center region when the navigation bar is the browser or editor modes. In navigation mode, these items will be placed in the overflow.

pub fn setCenterItemGroups( &self, center_item_groups: &NSArray<UIBarButtonItemGroup>, )

Setter for centerItemGroups.

This is copied when set.

pub fn trailingItemGroups(&self) -> Retained<NSArray<UIBarButtonItemGroup>>

UIBarButtonItemGroups to be displayed in the trailing section of the navigation bar. Items set via this method will replace items set via .rightBarButtonItem or .rightBarButtonItems.

pub fn setTrailingItemGroups( &self, trailing_item_groups: &NSArray<UIBarButtonItemGroup>, )

Setter for trailingItemGroups.

This is copied when set.

pub fn pinnedTrailingGroup(&self) -> Option<Retained<UIBarButtonItemGroup>>

Set a group of items to be placed on the trailing edge of the bar, past the overflow & search items (if present). If you wish to have more than 1 item, it is HIGHLY recommended to assign a representative item to the group to avoid overflowing the bar as this group cannot move to the overflow.

pub fn setPinnedTrailingGroup( &self, pinned_trailing_group: Option<&UIBarButtonItemGroup>, )

Setter for pinnedTrailingGroup.

pub fn additionalOverflowItems(&self) -> Option<Retained<UIDeferredMenuElement>>

Additional items to add to the overflow menu. Setting this property to a non-nil value will force the overflow button to appear, regardless of if you provide any content in the element’s callback. Items returned are displayed directly in the presented menu.

pub fn setAdditionalOverflowItems( &self, additional_overflow_items: Option<&UIDeferredMenuElement>, )

Setter for additionalOverflowItems.

pub fn overflowPresentationSource( &self, ) -> Option<Retained<ProtocolObject<dyn UIPopoverPresentationControllerSourceItem>>>

If the overflow button for this item is currently present, then this will return a non-nil item that can be used as a presentation source (e.g. for popovers). Otherwise it will return nil.

pub fn largeTitleDisplayMode(&self) -> UINavigationItemLargeTitleDisplayMode

When UINavigationBar.prefersLargeTitles=YES, this property controls when the larger out-of-line title is displayed. If prefersLargeTitles=NO, this property has no effect. The default value is Automatic.

pub fn setLargeTitleDisplayMode( &self, large_title_display_mode: UINavigationItemLargeTitleDisplayMode, )

Setter for largeTitleDisplayMode.

pub fn style(&self) -> UINavigationItemStyle

Controls how content defined by this item is laid out in the navigation bar.

pub fn setStyle(&self, style: UINavigationItemStyle)

Setter for style.

pub fn searchController(&self) -> Option<Retained<UISearchController>>

A view controller that will be shown inside of a navigation controller can assign a UISearchController to this property to display the search controller’s search bar in its containing navigation controller’s navigation bar.

pub fn setSearchController( &self, search_controller: Option<&UISearchController>, )

Setter for searchController.

pub fn hidesSearchBarWhenScrolling(&self) -> bool

If this property is true (the default), the searchController’s search bar will hide as the user scrolls in the top view controller’s scroll view. If false, the search bar will remain visible and pinned underneath the navigation bar. Not applicable and ignored for UINavigationItemSearchBarPlacementIntegrated

pub fn setHidesSearchBarWhenScrolling( &self, hides_search_bar_when_scrolling: bool, )

Setter for hidesSearchBarWhenScrolling.

pub fn preferredSearchBarPlacement(&self) -> UINavigationItemSearchBarPlacement

The preferred search bar placement, when a search controller is assigned to this item.

pub fn setPreferredSearchBarPlacement( &self, preferred_search_bar_placement: UINavigationItemSearchBarPlacement, )

Setter for preferredSearchBarPlacement.

pub fn searchBarPlacement(&self) -> UINavigationItemSearchBarPlacement

The realized placement. Only valid if a search controller has been assigned to this item.

pub fn searchBarPlacementBarButtonItem(&self) -> Retained<UIBarButtonItem>

When searchBarPlacement is .integrated or .integratedButton and a search controller is present, use this bar button item in the view controller’s toolbarItems to control the placement of the search bar among them when the search bar is appearing in the UIToolbar on iPhone. Without this bar button item, the positioning for the search bar defaults to trailingmost for the UIToolbar case. This bar button item will be ignored during toolbar layout if searchController is nil. UIBarButtonItemGroup will throw an NSInvalidArgumentException when this bar button item is included in its initialization. UINavigationItem will throw an NSInvalidArgumentException when this bar button item is included in leftBarButtonItems or rightBarButtonItems.

pub fn searchBarPlacementAllowsToolbarIntegration(&self) -> bool

Defaults to YES Set to NO to prevent the search bar from being placed among other UIToolbar items on iPhone

pub fn setSearchBarPlacementAllowsToolbarIntegration( &self, search_bar_placement_allows_toolbar_integration: bool, )

Setter for searchBarPlacementAllowsToolbarIntegration.

pub fn searchBarPlacementAllowsExternalIntegration(&self) -> bool

AllowsExternalIntegration means that something at a higher scope may take the search bar and integrate it somewhere other than the navigation bar (or toolbar) directly associated with this navigation item. On iOS 26, UISplitViewController uses this to allow Mac-like placement of the search bar on iPad. Defaults to NO Set to YES to allow the containing UISplitViewController to place the search bar in the navigation bar for the trailingmost column independent of the column used for this view controller. Ignored when searchController is nil, when searchBarPlacement is .stacked, or when not contained in a UISplitViewController. When the top view controller’s navigation item has this property set to YES in more than one column at the same time, only one will be respected; the columns are checked in order of .inspector, .secondary, .supplementary, .primary If the search bar has a scope bar, the scope bar will not be moved into a different column.

pub fn setSearchBarPlacementAllowsExternalIntegration( &self, search_bar_placement_allows_external_integration: bool, )

Setter for searchBarPlacementAllowsExternalIntegration.

pub fn standardAppearance(&self) -> Option<Retained<UINavigationBarAppearance>>

When set and this item is topmost, overrides the hosting navigation bar’s standardAppearance. See UINavigationBar.standardAppearance for further details.

pub fn setStandardAppearance( &self, standard_appearance: Option<&UINavigationBarAppearance>, )

Setter for standardAppearance.

This is copied when set.

pub fn compactAppearance(&self) -> Option<Retained<UINavigationBarAppearance>>

When set and this item is topmost, overrides the hosting navigation bar’s compactAppearance. See UINavigationBar.compactAppearance for further details.

pub fn setCompactAppearance( &self, compact_appearance: Option<&UINavigationBarAppearance>, )

Setter for compactAppearance.

This is copied when set.

pub fn scrollEdgeAppearance( &self, ) -> Option<Retained<UINavigationBarAppearance>>

When set and this item is topmost, overrides the hosting navigation bar’s scrollEdgeAppearance. See UINavigationBar.scrollEdgeAppearance for further details.

pub fn setScrollEdgeAppearance( &self, scroll_edge_appearance: Option<&UINavigationBarAppearance>, )

Setter for scrollEdgeAppearance.

This is copied when set.

pub fn compactScrollEdgeAppearance( &self, ) -> Option<Retained<UINavigationBarAppearance>>

When set and this item is topmost, overrides the hosting navigation bar’s compactScrollEdgeAppearance. See UINavigationBar.h for further details.

pub fn setCompactScrollEdgeAppearance( &self, compact_scroll_edge_appearance: Option<&UINavigationBarAppearance>, )

Setter for compactScrollEdgeAppearance.

This is copied when set.

impl UINavigationItem

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new(mtm: MainThreadMarker) -> Retained<Self>