objc2_ui_kit

Struct UIPasteboard

pub struct UIPasteboard { /* private fields */ }

Appleโs documentation

Implementations

impl UIPasteboard

pub fn generalPasteboard() -> Retained<UIPasteboard>

pub fn pasteboardWithName_create( pasteboard_name: &UIPasteboardName, create: bool, ) -> Option<Retained<UIPasteboard>>

pub fn pasteboardWithUniqueName() -> Retained<UIPasteboard>

pub unsafe fn name(&self) -> Retained<UIPasteboardName>

This property is not atomic.

Safety

This might not be thread-safe.

pub fn removePasteboardWithName(pasteboard_name: &UIPasteboardName)

pub unsafe fn isPersistent(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub fn setPersistent(&self, persistent: bool)

๐Deprecated: Do not set persistence on pasteboards. This property is set automatically.

pub unsafe fn changeCount(&self) -> NSInteger

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn itemProviders(&self) -> Retained<NSArray<NSItemProvider>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setItemProviders(&self, item_providers: &NSArray<NSItemProvider>)

Setter for itemProviders.

This is copied when set.

Safety

This might not be thread-safe.

pub fn setItemProviders_localOnly_expirationDate( &self, item_providers: &NSArray<NSItemProvider>, local_only: bool, expiration_date: Option<&NSDate>, )

pub fn setObjects( &self, objects: &NSArray<ProtocolObject<dyn NSItemProviderWriting>>, )

pub fn setObjects_localOnly_expirationDate( &self, objects: &NSArray<ProtocolObject<dyn NSItemProviderWriting>>, local_only: bool, expiration_date: Option<&NSDate>, )

pub unsafe fn pasteboardTypes(&self) -> Retained<NSArray<NSString>>

This property is not atomic.

Safety

This might not be thread-safe.

pub fn containsPasteboardTypes( &self, pasteboard_types: &NSArray<NSString>, ) -> bool

pub fn dataForPasteboardType( &self, pasteboard_type: &NSString, ) -> Option<Retained<NSData>>

pub fn valueForPasteboardType( &self, pasteboard_type: &NSString, ) -> Option<Retained<AnyObject>>

pub unsafe fn setValue_forPasteboardType( &self, value: &AnyObject, pasteboard_type: &NSString, )

Safety

value should be of the correct type.

pub fn setData_forPasteboardType( &self, data: &NSData, pasteboard_type: &NSString, )

pub unsafe fn numberOfItems(&self) -> NSInteger

This property is not atomic.

Safety

This might not be thread-safe.

pub fn pasteboardTypesForItemSet( &self, item_set: Option<&NSIndexSet>, ) -> Option<Retained<NSArray<NSArray<NSString>>>>

pub fn containsPasteboardTypes_inItemSet( &self, pasteboard_types: &NSArray<NSString>, item_set: Option<&NSIndexSet>, ) -> bool

pub fn itemSetWithPasteboardTypes( &self, pasteboard_types: &NSArray<NSString>, ) -> Option<Retained<NSIndexSet>>

pub fn valuesForPasteboardType_inItemSet( &self, pasteboard_type: &NSString, item_set: Option<&NSIndexSet>, ) -> Option<Retained<NSArray>>

pub fn dataForPasteboardType_inItemSet( &self, pasteboard_type: &NSString, item_set: Option<&NSIndexSet>, ) -> Option<Retained<NSArray<NSData>>>

pub unsafe fn items( &self, ) -> Retained<NSArray<NSDictionary<NSString, AnyObject>>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setItems( &self, items: &NSArray<NSDictionary<NSString, AnyObject>>, )

Setter for items.

This is copied when set.

Safety

items generic generic should be of the correct type.

This might not be thread-safe.

pub unsafe fn addItems( &self, items: &NSArray<NSDictionary<NSString, AnyObject>>, )

Safety

items generic generic should be of the correct type.

pub unsafe fn setItems_options( &self, items: &NSArray<NSDictionary<NSString, AnyObject>>, options: &NSDictionary<UIPasteboardOption, AnyObject>, )

Safety

items generic generic should be of the correct type.

options generic should be of the correct type.

pub unsafe fn string(&self) -> Option<Retained<NSString>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setString(&self, string: Option<&NSString>)

Setter for string.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn strings(&self) -> Option<Retained<NSArray<NSString>>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setStrings(&self, strings: Option<&NSArray<NSString>>)

Setter for strings.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn URL(&self) -> Option<Retained<NSURL>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setURL(&self, url: Option<&NSURL>)

Setter for URL.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn URLs(&self) -> Option<Retained<NSArray<NSURL>>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setURLs(&self, ur_ls: Option<&NSArray<NSURL>>)

Setter for URLs.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn image(&self) -> Option<Retained<UIImage>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setImage(&self, image: Option<&UIImage>)

Setter for image.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn images(&self) -> Option<Retained<NSArray<UIImage>>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setImages(&self, images: Option<&NSArray<UIImage>>)

Setter for images.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn color(&self) -> Option<Retained<UIColor>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setColor(&self, color: Option<&UIColor>)

Setter for color.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn colors(&self) -> Option<Retained<NSArray<UIColor>>>

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn setColors(&self, colors: Option<&NSArray<UIColor>>)

Setter for colors.

This is copied when set.

Safety

This might not be thread-safe.

pub unsafe fn hasStrings(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn hasURLs(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn hasImages(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn hasColors(&self) -> bool

This property is not atomic.

Safety

This might not be thread-safe.

pub fn detectPatternsForPatterns_completionHandler( &self, patterns: &NSSet<UIPasteboardDetectionPattern>, completion_handler: &DynBlock<dyn Fn(*mut NSSet<UIPasteboardDetectionPattern>, *mut NSError)>, )

Detects patterns in the first pasteboard item.

Parameter patterns: Detect only these patterns.

Parameter completionHandler: Receives which patterns were detected, or an error.

pub fn detectPatternsForPatterns_inItemSet_completionHandler( &self, patterns: &NSSet<UIPasteboardDetectionPattern>, item_set: Option<&NSIndexSet>, completion_handler: &DynBlock<dyn Fn(*mut NSArray<NSSet<UIPasteboardDetectionPattern>>, *mut NSError)>, )

Detects patterns in the specified pasteboard items.

Parameter patterns: Detect only these patterns.

Parameter itemSet: Specifies which pasteboard items by their position. Nil means all items.

Parameter completionHandler: Receives which patterns were detected per item specified, or an error.

pub fn detectValuesForPatterns_completionHandler( &self, patterns: &NSSet<UIPasteboardDetectionPattern>, completion_handler: &DynBlock<dyn Fn(*mut NSDictionary<UIPasteboardDetectionPattern, AnyObject>, *mut NSError)>, )

Detects patterns and corresponding values in the first pasteboard item.

Parameter patterns: Detect only these patterns.

Parameter completionHandler: Receives which patterns and values were detected, or an error.

pub fn detectValuesForPatterns_inItemSet_completionHandler( &self, patterns: &NSSet<UIPasteboardDetectionPattern>, item_set: Option<&NSIndexSet>, completion_handler: &DynBlock<dyn Fn(*mut NSArray<NSDictionary<UIPasteboardDetectionPattern, AnyObject>>, *mut NSError)>, )

Detects patterns and corresponding values in the specified pasteboard items.

Parameter patterns: Detect only these patterns.

Parameter itemSet: Specifies which pasteboard items by their position. Nil means all items.

Parameter completionHandler: Receives which patterns and values were detected per item specified, or an error.

impl UIPasteboard

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new() -> Retained<Self>