objc2_ui_kit

Struct UIResponder

pub struct UIResponder { /* private fields */ }

Apple’s documentation

Implementations

impl UIResponder

pub fn nextResponder(&self) -> Option<Retained<UIResponder>>

pub fn canBecomeFirstResponder(&self) -> bool

pub fn becomeFirstResponder(&self) -> bool

pub fn canResignFirstResponder(&self) -> bool

pub fn resignFirstResponder(&self) -> bool

pub fn isFirstResponder(&self) -> bool

pub fn touchesBegan_withEvent( &self, touches: &NSSet<UITouch>, event: Option<&UIEvent>, )

pub fn touchesMoved_withEvent( &self, touches: &NSSet<UITouch>, event: Option<&UIEvent>, )

pub fn touchesEnded_withEvent( &self, touches: &NSSet<UITouch>, event: Option<&UIEvent>, )

pub fn touchesCancelled_withEvent( &self, touches: &NSSet<UITouch>, event: Option<&UIEvent>, )

pub fn touchesEstimatedPropertiesUpdated(&self, touches: &NSSet<UITouch>)

pub fn pressesBegan_withEvent( &self, presses: &NSSet<UIPress>, event: Option<&UIPressesEvent>, )

pub fn pressesChanged_withEvent( &self, presses: &NSSet<UIPress>, event: Option<&UIPressesEvent>, )

pub fn pressesEnded_withEvent( &self, presses: &NSSet<UIPress>, event: Option<&UIPressesEvent>, )

pub fn pressesCancelled_withEvent( &self, presses: &NSSet<UIPress>, event: Option<&UIPressesEvent>, )

pub fn motionBegan_withEvent( &self, motion: UIEventSubtype, event: Option<&UIEvent>, )

pub fn motionEnded_withEvent( &self, motion: UIEventSubtype, event: Option<&UIEvent>, )

pub fn motionCancelled_withEvent( &self, motion: UIEventSubtype, event: Option<&UIEvent>, )

pub fn remoteControlReceivedWithEvent(&self, event: Option<&UIEvent>)

pub unsafe fn canPerformAction_withSender( &self, action: Sel, sender: Option<&AnyObject>, ) -> bool

Safety

action must be a valid selector.

sender should be of the correct type.

pub unsafe fn targetForAction_withSender( &self, action: Sel, sender: Option<&AnyObject>, ) -> Option<Retained<AnyObject>>

Safety

action must be a valid selector.

sender should be of the correct type.

pub fn buildMenuWithBuilder(&self, builder: &ProtocolObject<dyn UIMenuBuilder>)

pub fn validateCommand(&self, command: &UICommand)

pub fn providerForDeferredMenuElement( &self, deferred_element: &UIDeferredMenuElement, ) -> Option<Retained<UIDeferredMenuElementProvider>>

Asks the responder for an element provider to fulfill the given focus-based deferred element. Check the identifier of the deferred element to identify which deferred element this is. By default, this returns nil. Return a non-nil provider to make this responder responsible for providing elements for this fulfillment of the deferred element.

pub fn undoManager(&self) -> Option<Retained<NSUndoManager>>

pub fn editingInteractionConfiguration( &self, ) -> UIEditingInteractionConfiguration

impl UIResponder

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new(mtm: MainThreadMarker) -> Retained<Self>

impl UIResponder

UIResponderKeyCommands.

pub fn keyCommands(&self) -> Option<Retained<NSArray<UIKeyCommand>>>

impl UIResponder

UIResponderInputViewAdditions.

pub fn inputView(&self) -> Option<Retained<UIView>>

pub fn inputAccessoryView(&self) -> Option<Retained<UIView>>

pub fn inputAssistantItem(&self) -> Retained<UITextInputAssistantItem>

This method is for clients that wish to put buttons on the Shortcuts Bar, shown on top of the keyboard. You may modify the returned inputAssistantItem to add to or replace the existing items on the bar. Modifications made to the returned UITextInputAssistantItem are reflected automatically. This method should not be overridden. Goes up the responder chain.

pub fn inputViewController(&self) -> Option<Retained<UIInputViewController>>

pub fn inputAccessoryViewController( &self, ) -> Option<Retained<UIInputViewController>>

pub fn textInputMode(&self) -> Option<Retained<UITextInputMode>>

pub fn textInputContextIdentifier(&self) -> Option<Retained<NSString>>

pub fn clearTextInputContextIdentifier( identifier: &NSString, mtm: MainThreadMarker, )

pub fn reloadInputViews(&self)

impl UIResponder

ActivityContinuation.

pub fn userActivity(&self) -> Option<Retained<NSUserActivity>>

pub fn setUserActivity(&self, user_activity: Option<&NSUserActivity>)

Setter for userActivity.

pub fn updateUserActivityState(&self, activity: &NSUserActivity)

pub fn restoreUserActivityState(&self, activity: &NSUserActivity)

impl UIResponder

This impl block contains no items.

UIPasteConfigurationSupporting.

impl UIResponder

UICaptureTextFromCameraSupporting.

pub unsafe fn captureTextFromCamera(&self, sender: Option<&AnyObject>)

Safety

sender should be of the correct type.

impl UIResponder

UIActivityItemsConfiguration.

pub fn activityItemsConfiguration( &self, ) -> Option<Retained<ProtocolObject<dyn UIActivityItemsConfigurationReading>>>

pub fn setActivityItemsConfiguration( &self, activity_items_configuration: Option<&ProtocolObject<dyn UIActivityItemsConfigurationReading>>, )

Setter for activityItemsConfiguration.