objc2_ui_kit

Struct UITextField

pub struct UITextField { /* private fields */ }

Apple’s documentation

Implementations

impl UITextField

pub fn text(&self) -> Option<Retained<NSString>>

pub fn setText(&self, text: Option<&NSString>)

Setter for text.

This is copied when set.

pub fn attributedText(&self) -> Option<Retained<NSAttributedString>>

pub fn setAttributedText(&self, attributed_text: Option<&NSAttributedString>)

Setter for attributedText.

This is copied when set.

pub fn textColor(&self) -> Option<Retained<UIColor>>

pub fn setTextColor(&self, text_color: Option<&UIColor>)

Setter for textColor.

pub fn font(&self) -> Option<Retained<UIFont>>

pub fn setFont(&self, font: Option<&UIFont>)

Setter for font.

pub fn textAlignment(&self) -> NSTextAlignment

pub fn setTextAlignment(&self, text_alignment: NSTextAlignment)

Setter for textAlignment.

pub fn borderStyle(&self) -> UITextBorderStyle

pub fn setBorderStyle(&self, border_style: UITextBorderStyle)

Setter for borderStyle.

pub fn defaultTextAttributes( &self, ) -> Retained<NSDictionary<NSAttributedStringKey, AnyObject>>

pub unsafe fn setDefaultTextAttributes( &self, default_text_attributes: &NSDictionary<NSAttributedStringKey, AnyObject>, )

Setter for defaultTextAttributes.

This is copied when set.

Safety

default_text_attributes generic should be of the correct type.

pub fn placeholder(&self) -> Option<Retained<NSString>>

pub fn setPlaceholder(&self, placeholder: Option<&NSString>)

Setter for placeholder.

This is copied when set.

pub fn attributedPlaceholder(&self) -> Option<Retained<NSAttributedString>>

pub fn setAttributedPlaceholder( &self, attributed_placeholder: Option<&NSAttributedString>, )

Setter for attributedPlaceholder.

This is copied when set.

pub fn clearsOnBeginEditing(&self) -> bool

pub fn setClearsOnBeginEditing(&self, clears_on_begin_editing: bool)

Setter for clearsOnBeginEditing.

pub fn adjustsFontSizeToFitWidth(&self) -> bool

pub fn setAdjustsFontSizeToFitWidth(&self, adjusts_font_size_to_fit_width: bool)

Setter for adjustsFontSizeToFitWidth.

pub fn minimumFontSize(&self) -> CGFloat

pub fn setMinimumFontSize(&self, minimum_font_size: CGFloat)

Setter for minimumFontSize.

pub fn delegate( &self, ) -> Option<Retained<ProtocolObject<dyn UITextFieldDelegate>>>

pub fn setDelegate( &self, delegate: Option<&ProtocolObject<dyn UITextFieldDelegate>>, )

Setter for delegate.

This is a [weak property][objc2::topics::weak_property].

pub fn background(&self) -> Option<Retained<UIImage>>

pub fn setBackground(&self, background: Option<&UIImage>)

Setter for background.

pub fn disabledBackground(&self) -> Option<Retained<UIImage>>

pub fn setDisabledBackground(&self, disabled_background: Option<&UIImage>)

Setter for disabledBackground.

pub fn isEditing(&self) -> bool

pub fn allowsEditingTextAttributes(&self) -> bool

pub fn setAllowsEditingTextAttributes( &self, allows_editing_text_attributes: bool, )

Setter for allowsEditingTextAttributes.

pub fn typingAttributes( &self, ) -> Option<Retained<NSDictionary<NSAttributedStringKey, AnyObject>>>

pub unsafe fn setTypingAttributes( &self, typing_attributes: Option<&NSDictionary<NSAttributedStringKey, AnyObject>>, )

Setter for typingAttributes.

This is copied when set.

Safety

typing_attributes generic should be of the correct type.

pub fn clearButtonMode(&self) -> UITextFieldViewMode

pub fn setClearButtonMode(&self, clear_button_mode: UITextFieldViewMode)

Setter for clearButtonMode.

pub fn leftView(&self) -> Option<Retained<UIView>>

pub fn setLeftView(&self, left_view: Option<&UIView>)

Setter for leftView.

pub fn leftViewMode(&self) -> UITextFieldViewMode

pub fn setLeftViewMode(&self, left_view_mode: UITextFieldViewMode)

Setter for leftViewMode.

pub fn rightView(&self) -> Option<Retained<UIView>>

pub fn setRightView(&self, right_view: Option<&UIView>)

Setter for rightView.

pub fn rightViewMode(&self) -> UITextFieldViewMode

pub fn setRightViewMode(&self, right_view_mode: UITextFieldViewMode)

Setter for rightViewMode.

pub fn borderRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn textRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn placeholderRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn editingRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn clearButtonRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn leftViewRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn rightViewRectForBounds(&self, bounds: CGRect) -> CGRect

pub fn drawTextInRect(&self, rect: CGRect)

👎Deprecated: This method is no longer called.

pub fn drawPlaceholderInRect(&self, rect: CGRect)

pub fn inputView(&self) -> Option<Retained<UIView>>

pub fn setInputView(&self, input_view: Option<&UIView>)

Setter for inputView.

pub fn inputAccessoryView(&self) -> Option<Retained<UIView>>

pub fn setInputAccessoryView(&self, input_accessory_view: Option<&UIView>)

Setter for inputAccessoryView.

pub fn clearsOnInsertion(&self) -> bool

pub fn setClearsOnInsertion(&self, clears_on_insertion: bool)

Setter for clearsOnInsertion.

impl UITextField

Methods declared on superclass UIControl.

pub fn initWithFrame(this: Allocated<Self>, frame: CGRect) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

pub fn initWithFrame_primaryAction( this: Allocated<Self>, frame: CGRect, primary_action: Option<&UIAction>, ) -> Retained<Self>

Initializes the control and adds primaryAction for the UIControlEventPrimaryActionTriggered control event. Subclasses of UIControl may alter or add behaviors around the usage of primaryAction, see subclass documentation of this initializer for additional information.

impl UITextField

Methods declared on superclass UIView.

pub fn init(this: Allocated<Self>) -> Retained<Self>

impl UITextField

Methods declared on superclass NSObject.

pub fn new(mtm: MainThreadMarker) -> Retained<Self>

impl UITextField

UIInteractionStateRestorable.

pub fn interactionState(&self) -> Retained<AnyObject>

pub unsafe fn setInteractionState(&self, interaction_state: &AnyObject)

Setter for interactionState.

This is copied when set.

Safety

interaction_state should be of the correct type.