objc2_ui_kit

Struct UIToolbar

pub struct UIToolbar { /* private fields */ }

Apple’s documentation

Implementations

impl UIToolbar

pub fn barStyle(&self) -> UIBarStyle

pub fn setBarStyle(&self, bar_style: UIBarStyle)

Setter for barStyle.

pub fn items(&self) -> Option<Retained<NSArray<UIBarButtonItem>>>

pub fn setItems(&self, items: Option<&NSArray<UIBarButtonItem>>)

Setter for items.

This is copied when set.

pub fn isTranslucent(&self) -> bool

pub fn setTranslucent(&self, translucent: bool)

Setter for isTranslucent.

pub fn setItems_animated( &self, items: Option<&NSArray<UIBarButtonItem>>, animated: bool, )

pub fn tintColor(&self) -> Option<Retained<UIColor>>

pub unsafe fn setTintColor(&self, tint_color: Option<&UIColor>)

Setter for tintColor.

Safety

tint_color might not allow None.

pub fn barTintColor(&self) -> Option<Retained<UIColor>>

pub fn setBarTintColor(&self, bar_tint_color: Option<&UIColor>)

Setter for barTintColor.

pub fn setBackgroundImage_forToolbarPosition_barMetrics( &self, background_image: Option<&UIImage>, top_or_bottom: UIBarPosition, bar_metrics: UIBarMetrics, )

pub fn backgroundImageForToolbarPosition_barMetrics( &self, top_or_bottom: UIBarPosition, bar_metrics: UIBarMetrics, ) -> Option<Retained<UIImage>>

pub fn setShadowImage_forToolbarPosition( &self, shadow_image: Option<&UIImage>, top_or_bottom: UIBarPosition, )

pub fn shadowImageForToolbarPosition( &self, top_or_bottom: UIBarPosition, ) -> Option<Retained<UIImage>>

pub fn standardAppearance(&self) -> Retained<UIToolbarAppearance>

Describes the appearance attributes for the toolbar to use when it is displayed with its standard height.

pub fn setStandardAppearance(&self, standard_appearance: &UIToolbarAppearance)

Setter for standardAppearance.

This is copied when set.

pub fn compactAppearance(&self) -> Option<Retained<UIToolbarAppearance>>

Describes the appearance attributes for the toolbar to use when it is displayed with its compact height. If not set, the standardAppearance will be used instead.

pub fn setCompactAppearance( &self, compact_appearance: Option<&UIToolbarAppearance>, )

Setter for compactAppearance.

This is copied when set.

pub fn scrollEdgeAppearance(&self) -> Option<Retained<UIToolbarAppearance>>

Describes the appearance attributes for the toolbar to use at standard height when an observable scroll view is scrolled to the bottom. If not set, standardAppearance will be used instead.

pub fn setScrollEdgeAppearance( &self, scroll_edge_appearance: Option<&UIToolbarAppearance>, )

Setter for scrollEdgeAppearance.

This is copied when set.

pub fn compactScrollEdgeAppearance( &self, ) -> Option<Retained<UIToolbarAppearance>>

Describes the appearance attributes for the toolbar to use at compact height when an observable scroll view is scrolled to the bottom. If not set, will fall back to scrollEdgeAppearance followed by compactAppearance and finally standardAppearance.

pub fn setCompactScrollEdgeAppearance( &self, compact_scroll_edge_appearance: Option<&UIToolbarAppearance>, )

Setter for compactScrollEdgeAppearance.

This is copied when set.

pub fn delegate( &self, ) -> Option<Retained<ProtocolObject<dyn UIToolbarDelegate>>>

pub fn setDelegate( &self, delegate: Option<&ProtocolObject<dyn UIToolbarDelegate>>, )

Setter for delegate.

This is a [weak property][objc2::topics::weak_property].

impl UIToolbar

Methods declared on superclass UIView.

pub fn initWithFrame(this: Allocated<Self>, frame: CGRect) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

pub fn init(this: Allocated<Self>) -> Retained<Self>

impl UIToolbar

Methods declared on superclass NSObject.

pub fn new(mtm: MainThreadMarker) -> Retained<Self>