objc2_ui_kit

Trait UIDocumentBrowserViewControllerDelegate

pub unsafe trait UIDocumentBrowserViewControllerDelegate: NSObjectProtocol {
    // Provided methods
    fn documentBrowser_didPickDocumentURLs(
        &self,
        controller: &UIDocumentBrowserViewController,
        document_ur_ls: &NSArray<NSURL>,
    )
       where Self: Sized + Message { ... }
    fn documentBrowser_didPickDocumentsAtURLs(
        &self,
        controller: &UIDocumentBrowserViewController,
        document_ur_ls: &NSArray<NSURL>,
    )
       where Self: Sized + Message { ... }
    fn documentBrowser_didRequestDocumentCreationWithHandler(
        &self,
        controller: &UIDocumentBrowserViewController,
        import_handler: &DynBlock<dyn Fn(*mut NSURL, UIDocumentBrowserImportMode)>,
    )
       where Self: Sized + Message { ... }
    fn documentBrowser_didImportDocumentAtURL_toDestinationURL(
        &self,
        controller: &UIDocumentBrowserViewController,
        source_url: &NSURL,
        destination_url: &NSURL,
    )
       where Self: Sized + Message { ... }
    fn documentBrowser_failedToImportDocumentAtURL_error(
        &self,
        controller: &UIDocumentBrowserViewController,
        document_url: &NSURL,
        error: Option<&NSError>,
    )
       where Self: Sized + Message { ... }
    fn documentBrowser_applicationActivitiesForDocumentURLs(
        &self,
        controller: &UIDocumentBrowserViewController,
        document_ur_ls: &NSArray<NSURL>,
    ) -> Retained<NSArray<UIActivity>>
       where Self: Sized + Message { ... }
    fn documentBrowser_willPresentActivityViewController(
        &self,
        controller: &UIDocumentBrowserViewController,
        activity_view_controller: &UIActivityViewController,
    )
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn documentBrowser_didPickDocumentURLs( &self, controller: &UIDocumentBrowserViewController, document_ur_ls: &NSArray<NSURL>, )
where Self: Sized + Message,

👎Deprecated

Called when the user validates a selection of items to open or pick. If you have created this document manager to open files, you should then open the document, potentially using a UIDocumentBrowserTransitionControllerfor the transition.

fn documentBrowser_didPickDocumentsAtURLs( &self, controller: &UIDocumentBrowserViewController, document_ur_ls: &NSArray<NSURL>, )
where Self: Sized + Message,

fn documentBrowser_didRequestDocumentCreationWithHandler( &self, controller: &UIDocumentBrowserViewController, import_handler: &DynBlock<dyn Fn(*mut NSURL, UIDocumentBrowserImportMode)>, )
where Self: Sized + Message,

When the user requests the creation of a new document, this method will be called. The application can then, optionally, present UI it deems appropriate to let the user configure the new document (for example, it could show a list of templates). When done, create an empty document or a copy of your template to a temporary location. Then use the importHandler to pass the document URL to be imported. If the user cancels the document creation, it’s still expetected to call importHandler(nil, UIDocumentBrowserImportModeNone). The Document Browser will asynchronously move the document to its final destination and call back one of the following delegate methods: -documentBrowser:didImportDocumentURL:toDestinationURL: -documentBrowser:failedToImportDocumentAtURL:error: The app should wait for the delegate callbacks before presenting the imported document at the destination URL. Usage of UIDocument is recommended. If you do not implement this method, document creation will not be available.

fn documentBrowser_didImportDocumentAtURL_toDestinationURL( &self, controller: &UIDocumentBrowserViewController, source_url: &NSURL, destination_url: &NSURL, )
where Self: Sized + Message,

fn documentBrowser_failedToImportDocumentAtURL_error( &self, controller: &UIDocumentBrowserViewController, document_url: &NSURL, error: Option<&NSError>, )
where Self: Sized + Message,

fn documentBrowser_applicationActivitiesForDocumentURLs( &self, controller: &UIDocumentBrowserViewController, document_ur_ls: &NSArray<NSURL>, ) -> Retained<NSArray<UIActivity>>
where Self: Sized + Message,

Allows clients to add application specific UIActivity instances

fn documentBrowser_willPresentActivityViewController( &self, controller: &UIDocumentBrowserViewController, activity_view_controller: &UIActivityViewController, )
where Self: Sized + Message,

Implement this to customize the UIActivityViewController before it’s presented.