objc2_ui_kit

Trait UIDynamicItem

pub unsafe trait UIDynamicItem: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn center(&self) -> CGPoint
       where Self: Sized + Message { ... }
    fn setCenter(&self, center: CGPoint)
       where Self: Sized + Message { ... }
    fn bounds(&self) -> CGRect
       where Self: Sized + Message { ... }
    fn transform(&self) -> CGAffineTransform
       where Self: Sized + Message { ... }
    fn setTransform(&self, transform: CGAffineTransform)
       where Self: Sized + Message { ... }
    fn collisionBoundsType(&self) -> UIDynamicItemCollisionBoundsType
       where Self: Sized + Message { ... }
    fn collisionBoundingPath(&self) -> Retained<UIBezierPath>
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn center(&self) -> CGPoint
where Self: Sized + Message,

fn setCenter(&self, center: CGPoint)
where Self: Sized + Message,

Setter for center.

fn bounds(&self) -> CGRect
where Self: Sized + Message,

fn transform(&self) -> CGAffineTransform
where Self: Sized + Message,

fn setTransform(&self, transform: CGAffineTransform)
where Self: Sized + Message,

Setter for transform.

fn collisionBoundsType(&self) -> UIDynamicItemCollisionBoundsType
where Self: Sized + Message,

The collision type represents how the dynamics system will evaluate collisions with respect to the dynamic item. defaults to UIDynamicItemCollisionBoundsTypeRectangle

fn collisionBoundingPath(&self) -> Retained<UIBezierPath>
where Self: Sized + Message,

The path must represent a convex polygon with counter clockwise winding and no self intersection. The point (0,0) in the path corresponds to the dynamic item’s center.