objc2_ui_kit

Trait UIFocusEnvironment

pub unsafe trait UIFocusEnvironment: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn preferredFocusEnvironments(
        &self,
    ) -> Retained<NSArray<ProtocolObject<dyn UIFocusEnvironment>>>
       where Self: Sized + Message { ... }
    fn parentFocusEnvironment(
        &self,
    ) -> Option<Retained<ProtocolObject<dyn UIFocusEnvironment>>>
       where Self: Sized + Message { ... }
    fn focusItemContainer(
        &self,
    ) -> Option<Retained<ProtocolObject<dyn UIFocusItemContainer>>>
       where Self: Sized + Message { ... }
    fn setNeedsFocusUpdate(&self)
       where Self: Sized + Message { ... }
    fn updateFocusIfNeeded(&self)
       where Self: Sized + Message { ... }
    fn shouldUpdateFocusInContext(&self, context: &UIFocusUpdateContext) -> bool
       where Self: Sized + Message { ... }
    fn didUpdateFocusInContext_withAnimationCoordinator(
        &self,
        context: &UIFocusUpdateContext,
        coordinator: &UIFocusAnimationCoordinator,
    )
       where Self: Sized + Message { ... }
    fn soundIdentifierForFocusUpdateInContext(
        &self,
        context: &UIFocusUpdateContext,
    ) -> Option<Retained<UIFocusSoundIdentifier>>
       where Self: Sized + Message { ... }
    fn preferredFocusedView(&self) -> Option<Retained<UIView>>
       where Self: Sized + Message { ... }
    fn focusGroupIdentifier(&self) -> Option<Retained<NSString>>
       where Self: Sized + Message { ... }
}

Objects conforming to UIFocusEnvironment influence and respond to focus behavior within a specific area of the screen that they control.

See also Apple’s documentation

Provided Methods

fn preferredFocusEnvironments( &self, ) -> Retained<NSArray<ProtocolObject<dyn UIFocusEnvironment>>>
where Self: Sized + Message,

The preferred focus environments define where to search for the default focused item in an environment, such as when focus updates programmatically. Starting from the target environment, each preferred focus environment is recursively searched in the order of the array until an eligible, focusable item is found. Preferred focus environments can include focusable and non-focusable items, in addition to non-item environments. Returning an empty array is equivalent to returning an array containing only ‘self’.

fn parentFocusEnvironment( &self, ) -> Option<Retained<ProtocolObject<dyn UIFocusEnvironment>>>
where Self: Sized + Message,

The parent focus environment of this environment, or nil if no parent exists. NOTE: If you implement this method, you must return a non-nil value for parent focus environment, otherwise your focus environment will not participate in focus interactions.

fn focusItemContainer( &self, ) -> Option<Retained<ProtocolObject<dyn UIFocusItemContainer>>>
where Self: Sized + Message,

The container of any child focus items in this focus environment, or nil if no container exists.

fn setNeedsFocusUpdate(&self)
where Self: Sized + Message,

Marks this environment as needing a focus update, which if accepted will attempt to reset focus to this environment, or one of its preferred focus environments, on the next update cycle. If this environment does not currently contain the focused item, then calling this method has no effect. If a parent of this environment is also requesting focus, then this environment’s request is rejected in favor of the parent’s. NOTE: If you provide your own implementation, it must call [[UIFocusSystem focusSystemForEnvironment:self] requestFocusUpdateToEnvironment:self];

fn updateFocusIfNeeded(&self)
where Self: Sized + Message,

Forces focus to be updated immediately. If there is an environment that has requested a focus update via -setNeedsFocusUpdate, and the request was accepted, then focus will be updated to that environment or one of its preferred focus environments. NOTE: If you provide your own implementation, it must call [[UIFocusSystem focusSystemForEnvironment:self] updateFocusIfNeeded];.

fn shouldUpdateFocusInContext(&self, context: &UIFocusUpdateContext) -> bool
where Self: Sized + Message,

Asks whether the system should allow a focus update to occur.

fn didUpdateFocusInContext_withAnimationCoordinator( &self, context: &UIFocusUpdateContext, coordinator: &UIFocusAnimationCoordinator, )
where Self: Sized + Message,

Called when the screen’s focused item has been updated to a new item. Use the animation coordinator to schedule focus-related animations in response to the update.

fn soundIdentifierForFocusUpdateInContext( &self, context: &UIFocusUpdateContext, ) -> Option<Retained<UIFocusSoundIdentifier>>
where Self: Sized + Message,

Specifies an identifier corresponding to a sound that should be played for a focus update. Return UIFocusSoundIdentifierNone to opt out of sounds, UIFocusSoundIdentifierDefault for the system default sounds, a previously registered identifier for a custom sound, or nil to defer the decision to the parent.

fn preferredFocusedView(&self) -> Option<Retained<UIView>>
where Self: Sized + Message,

👎Deprecated

fn focusGroupIdentifier(&self) -> Option<Retained<NSString>>
where Self: Sized + Message,

The identifier of the focus group that this view belongs to. If this is nil, subviews inherit their superview’s focus group.