objc2_ui_kit

Trait UIFocusItem

pub unsafe trait UIFocusItem: UIFocusEnvironment + MainThreadOnly {
    // Provided methods
    fn canBecomeFocused(&self) -> bool
       where Self: Sized + Message { ... }
    fn frame(&self) -> CGRect
       where Self: Sized + Message { ... }
    fn focusEffect(&self) -> Option<Retained<UIFocusEffect>>
       where Self: Sized + Message { ... }
    fn focusGroupPriority(&self) -> UIFocusGroupPriority
       where Self: Sized + Message { ... }
    fn focusItemDeferralMode(&self) -> UIFocusItemDeferralMode
       where Self: Sized + Message { ... }
    fn isTransparentFocusItem(&self) -> bool
       where Self: Sized + Message { ... }
    fn didHintFocusMovement(&self, hint: &UIFocusMovementHint)
       where Self: Sized + Message { ... }
}

Objects conforming to UIFocusItem are considered capable of participating in focus. Only UIFocusItems can ever be focused.

See also Apple’s documentation

Provided Methods

fn canBecomeFocused(&self) -> bool
where Self: Sized + Message,

Indicates whether or not this item is currently allowed to become focused. Returning NO restricts the item from being focusable, even if it is visible in the user interface. For example, UIControls return NO if they are disabled.

fn frame(&self) -> CGRect
where Self: Sized + Message,

The geometric frame of this item, represented in the coordinateSpace of the UIFocusItemContainer in which it is contained.

fn focusEffect(&self) -> Option<Retained<UIFocusEffect>>
where Self: Sized + Message,

Describes a visual effect to apply when this item is focused. When not implemented, the system may create a default effect for this item. Returning nil indicates that the system should not apply any visual effects, and that the app will handle applying the appropriate visuals.

fn focusGroupPriority(&self) -> UIFocusGroupPriority
where Self: Sized + Message,

The priority this item has in its focus group. The higher the priority, the more likely it is to get picked when focus moves into this group. Note: this method can only be used to increase an item’s priority, not decrease it. For example if an item is currently selected, the actual priority of this item will be determined by MAX(focusGroupPriority, UIFocusGroupPrioritySelected).

fn focusItemDeferralMode(&self) -> UIFocusItemDeferralMode
where Self: Sized + Message,

If this property is present and returns UIFocusItemDeferralModeNever, the focus deferral will not be enabled again after the user engagement timeout has expired if this item is currently focused and programmatic focus updates pointing to this item will be executed immediatly. If it returns UIFocusItemDeferralModeAlways focus will always be deferred when this item is supposed to be focused. Does nothing when focus deferral is not supported on the platform.

fn isTransparentFocusItem(&self) -> bool
where Self: Sized + Message,

If this returns YES, the focus item is considered transparent in terms of occlusion. Items that are behind it are focusable. This value is ignored when the item is focusable, in which case the item is never considered transparent.

fn didHintFocusMovement(&self, hint: &UIFocusMovementHint)
where Self: Sized + Message,

Called whenever this focus item is hinting to the user a focus movement might occur. The provided object is mutated by the focus engine whenever the user’s finger moves.