objc2_ui_kit

Trait UIFocusItemContainer

pub unsafe trait UIFocusItemContainer: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn coordinateSpace(&self) -> Retained<ProtocolObject<dyn UICoordinateSpace>>
       where Self: Sized + Message { ... }
    fn focusItemsInRect(
        &self,
        rect: CGRect,
    ) -> Retained<NSArray<ProtocolObject<dyn UIFocusItem>>>
       where Self: Sized + Message { ... }
}

Objects conforming to UIFocusItemContainer are responsible for providing which focus items they contain and where they are.

See also Apple’s documentation

Provided Methods

fn coordinateSpace(&self) -> Retained<ProtocolObject<dyn UICoordinateSpace>>
where Self: Sized + Message,

The coordinate space of the focus items contained in this container. The focus items returned by focusItemsInRect: should report their frames in this coordinate space. If you are implementing this protocol, you may find it convenient to return the UIScreen as your coordinate space, and ensure that your contained items report their frames in screen space. Similarly, you might find that your focus items’ containing UIView or UIWindow is the most convenient coordinate space to use. You may also choose to implement your own object that conforms to UICoordinateSpace, if that is the most natural solution for your architecture.

fn focusItemsInRect( &self, rect: CGRect, ) -> Retained<NSArray<ProtocolObject<dyn UIFocusItem>>>
where Self: Sized + Message,

Returns an array of all focus items within this container that intersect with the provided rect. rect is expressed in coordinateSpace. Note: starting in iOS & tvOS 16.0, UIView will return its subviews from this method. If you override this method in a UIView subclass, it will be your responsibility to call super and merge your array of custom focus items with UIView’s default focus items.