objc2_ui_kit

Trait UIMenuLeaf

pub unsafe trait UIMenuLeaf: NSObjectProtocol + MainThreadOnly {

    // Provided methods
    fn title(&self) -> Retained<NSString>
       where Self: Sized + Message { ... }
    fn setTitle(&self, title: &NSString)
       where Self: Sized + Message { ... }
    fn image(&self) -> Option<Retained<UIImage>>
       where Self: Sized + Message { ... }
    fn setImage(&self, image: Option<&UIImage>)
       where Self: Sized + Message { ... }
    fn selectedImage(&self) -> Option<Retained<UIImage>>
       where Self: Sized + Message { ... }
    fn setSelectedImage(&self, selected_image: Option<&UIImage>)
       where Self: Sized + Message { ... }
    fn discoverabilityTitle(&self) -> Option<Retained<NSString>>
       where Self: Sized + Message { ... }
    fn setDiscoverabilityTitle(&self, discoverability_title: Option<&NSString>)
       where Self: Sized + Message { ... }
    fn attributes(&self) -> UIMenuElementAttributes
       where Self: Sized + Message { ... }
    fn setAttributes(&self, attributes: UIMenuElementAttributes)
       where Self: Sized + Message { ... }
    fn state(&self) -> UIMenuElementState
       where Self: Sized + Message { ... }
    fn setState(&self, state: UIMenuElementState)
       where Self: Sized + Message { ... }
    fn repeatBehavior(&self) -> UIMenuElementRepeatBehavior
       where Self: Sized + Message { ... }
    fn setRepeatBehavior(&self, repeat_behavior: UIMenuElementRepeatBehavior)
       where Self: Sized + Message { ... }
    fn sender(&self) -> Option<Retained<AnyObject>>
       where Self: Sized + Message { ... }
    fn presentationSourceItem(
        &self,
    ) -> Option<Retained<ProtocolObject<dyn UIPopoverPresentationControllerSourceItem>>>
       where Self: Sized + Message { ... }
    unsafe fn performWithSender_target(
        &self,
        sender: Option<&AnyObject>,
        target: Option<&AnyObject>,
    )
       where Self: Sized + Message { ... }

}

Apple’s documentation

Provided Methods

fn title(&self) -> Retained<NSString>
where Self: Sized + Message,

Short display title.

fn setTitle(&self, title: &NSString)
where Self: Sized + Message,

Setter for title.

This is copied when set.

fn image(&self) -> Option<Retained<UIImage>>
where Self: Sized + Message,

Image that can appear next to this element.

fn setImage(&self, image: Option<&UIImage>)
where Self: Sized + Message,

Setter for image.

This is copied when set.

fn selectedImage(&self) -> Option<Retained<UIImage>>
where Self: Sized + Message,

Image that can appear next to this action when the state is UIMenuElementStateOn

fn setSelectedImage(&self, selected_image: Option<&UIImage>)
where Self: Sized + Message,

Setter for selectedImage.

This is copied when set.

fn discoverabilityTitle(&self) -> Option<Retained<NSString>>
where Self: Sized + Message,

Elaborated title used in keyboard shortcut overlay.

fn setDiscoverabilityTitle(&self, discoverability_title: Option<&NSString>)
where Self: Sized + Message,

Setter for discoverabilityTitle.

This is copied when set.

fn attributes(&self) -> UIMenuElementAttributes
where Self: Sized + Message,

The element’s attributes.

fn setAttributes(&self, attributes: UIMenuElementAttributes)
where Self: Sized + Message,

Setter for attributes.

fn state(&self) -> UIMenuElementState
where Self: Sized + Message,

The element’s selection state.

fn setState(&self, state: UIMenuElementState)
where Self: Sized + Message,

Setter for state.

fn repeatBehavior(&self) -> UIMenuElementRepeatBehavior
where Self: Sized + Message,

The leaf’s preferred repeat behavior. Menu leaves can repeatedly perform their primary actions on prolonged interactions, such as by holding down their keyboard shortcut.

fn setRepeatBehavior(&self, repeat_behavior: UIMenuElementRepeatBehavior)
where Self: Sized + Message,

Setter for repeatBehavior.

fn sender(&self) -> Option<Retained<AnyObject>>
where Self: Sized + Message,

The object on behalf of which the element’s primary action is being performed. This property is only populated during the execution of the element’s primary action (i.e. its handler or selector) if available.

fn presentationSourceItem( &self, ) -> Option<Retained<ProtocolObject<dyn UIPopoverPresentationControllerSourceItem>>>
where Self: Sized + Message,

Source item for subsequent presentations. For example, an element displayed in a menu attached to a UIButtonmay return the button so that a popover may be presented from it when the element is tapped. If available, this property is only populated during the execution of the element’s action (i.e. its handler or selector).

unsafe fn performWithSender_target( &self, sender: Option<&AnyObject>, target: Option<&AnyObject>, )
where Self: Sized + Message,

Performs the element’s primary action (i.e. it’s handler or selector).

Safety

sender should be of the correct type.

target should be of the correct type.