objc2_ui_kit

Trait UIScreenshotServiceDelegate

pub unsafe trait UIScreenshotServiceDelegate: NSObjectProtocol + MainThreadOnly {
    // Provided method
    fn screenshotService_generatePDFRepresentationWithCompletion(
        &self,
        screenshot_service: &UIScreenshotService,
        completion_handler: &DynBlock<dyn Fn(*mut NSData, NSInteger, CGRect)>,
    )
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn screenshotService_generatePDFRepresentationWithCompletion( &self, screenshot_service: &UIScreenshotService, completion_handler: &DynBlock<dyn Fn(*mut NSData, NSInteger, CGRect)>, )
where Self: Sized + Message,

The delegate method to send the PDF data to screenshots

The delegate is required to send the data via the completion handler

Parameter screenshotService: The screenshot service object associated per scene

Parameter completionHandler: The block to execute when the PDF data is ready. If no PDF data is available, data can be nil. The indexOfCurrentPage is zero-based, and is the index of the current page of the snapshotted PDF. The rectInCurrentPage is the rect in PDF coordinates with respect to the current page. If the receiver cannot provide the visible area, pass CGRectZero for rectInCurrentPage.