objc2_ui_kit

Trait UITabBarControllerSidebarDelegate

pub unsafe trait UITabBarControllerSidebarDelegate: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn tabBarController_sidebarVisibilityWillChange_animator(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        animator: &ProtocolObject<dyn UITabBarControllerSidebarAnimating>,
    )
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_itemForRequest(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        request: &UITabSidebarItemRequest,
    ) -> Retained<UITabSidebarItem>
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_updateItem(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        item: &UITabSidebarItem,
    )
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_willBeginDisplayingTab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        tab: &UITab,
    )
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_didEndDisplayingTab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        tab: &UITab,
    )
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_leadingSwipeActionsConfigurationForTab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        tab: &UITab,
    ) -> Option<Retained<UISwipeActionsConfiguration>>
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_trailingSwipeActionsConfigurationForTab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        tab: &UITab,
    ) -> Option<Retained<UISwipeActionsConfiguration>>
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_contextMenuConfigurationForTab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        tab: &UITab,
    ) -> Option<Retained<UIContextMenuConfiguration>>
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_itemsForBeginningDragSession_tab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        drag_session: &ProtocolObject<dyn UIDragSession>,
        tab: &UITab,
    ) -> Retained<NSArray<UIDragItem>>
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_itemsForAddingToDragSession_tab(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        drag_session: &ProtocolObject<dyn UIDragSession>,
        tab: &UITab,
    ) -> Retained<NSArray<UIDragItem>>
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_sidebarAction_group_operationForAcceptingItemsFromDropSession(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        sidebar_action: &UIAction,
        group: &UITabGroup,
        session: &ProtocolObject<dyn UIDropSession>,
    ) -> UIDropOperation
       where Self: Sized + Message { ... }
    fn tabBarController_sidebar_sidebarAction_group_acceptItemsFromDropSession(
        &self,
        tab_bar_controller: &UITabBarController,
        sidebar: &UITabBarControllerSidebar,
        sidebar_action: &UIAction,
        group: &UITabGroup,
        session: &ProtocolObject<dyn UIDropSession>,
    )
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn tabBarController_sidebarVisibilityWillChange_animator( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, animator: &ProtocolObject<dyn UITabBarControllerSidebarAnimating>, )
where Self: Sized + Message,

Notifies the delegate when the visibility of the sidebar is about to change when sidebar.isHidden changes. Add animations to the animator to run alongside the visibility update. Alongside animations and completions will run immediately if the sidebar visibility is changed without animation.

fn tabBarController_sidebar_itemForRequest( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, request: &UITabSidebarItemRequest, ) -> Retained<UITabSidebarItem>
where Self: Sized + Message,

Return a UITabSidebarItem for the specified item request. When created, the item will be preconfigured to the appropriate defaults for its given content. If this method is not implemented, a default sidebar item will be provided for the request.

fn tabBarController_sidebar_updateItem( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, item: &UITabSidebarItem, )
where Self: Sized + Message,

Called whenever the sidebar item’s configurationState changes or the item is reconfigured. The passed in item will accrue all modifications until the delegate requests for a new sidebar item from the delegate method tabBarController:sidebar:itemForRequest:

fn tabBarController_sidebar_willBeginDisplayingTab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, tab: &UITab, )
where Self: Sized + Message,

Notifies the delegate when the sidebar is about to display the row representing the specified tab

fn tabBarController_sidebar_didEndDisplayingTab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, tab: &UITab, )
where Self: Sized + Message,

Notifies the delegate when the sidebar has finished displaying the row representing the specified tab

fn tabBarController_sidebar_leadingSwipeActionsConfigurationForTab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, tab: &UITab, ) -> Option<Retained<UISwipeActionsConfiguration>>
where Self: Sized + Message,

Called when the sidebar is about to show leading swipe actions for the specified tab. Return either a concrete UISwipeActionsConfiguration or nil if the tab does not show swipe actions.

fn tabBarController_sidebar_trailingSwipeActionsConfigurationForTab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, tab: &UITab, ) -> Option<Retained<UISwipeActionsConfiguration>>
where Self: Sized + Message,

Called when the sidebar is about to show trailing swipe actions for a particular tab. Return either a UISwipeActionsConfiguration object or nil if this tab does not show swipe actions.

fn tabBarController_sidebar_contextMenuConfigurationForTab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, tab: &UITab, ) -> Option<Retained<UIContextMenuConfiguration>>
where Self: Sized + Message,

Called when the sidebar is about to display a context menu for the specified tab. Return either a concrete UIContextMenuConfiguration or nil if the tab does not show context menus.

fn tabBarController_sidebar_itemsForBeginningDragSession_tab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, drag_session: &ProtocolObject<dyn UIDragSession>, tab: &UITab, ) -> Retained<NSArray<UIDragItem>>
where Self: Sized + Message,

Called when a new drag session has begun in the sidebar from the specified tab. Return drag items if the specified tab can be dragged, or an empty array if no drags should begin. Note that if drag items are returned on tabs in groups that allow reordering, then tab reordering is disabled when the sidebar is not in editing.

fn tabBarController_sidebar_itemsForAddingToDragSession_tab( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, drag_session: &ProtocolObject<dyn UIDragSession>, tab: &UITab, ) -> Retained<NSArray<UIDragItem>>
where Self: Sized + Message,

Called when a new drag session is requesting items to add to the existing drag session in the sidebar from the specified tab. Return items if the specified tab can add to the drag session, or an empty array if nothing should be added.

fn tabBarController_sidebar_sidebarAction_group_operationForAcceptingItemsFromDropSession( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, sidebar_action: &UIAction, group: &UITabGroup, session: &ProtocolObject<dyn UIDropSession>, ) -> UIDropOperation
where Self: Sized + Message,

Determines if items from the specified drop session can be dropped into the specified sidebarAction. If the operation is either a .move or .copy, then the drop will proceed and tabBarController:sidebar:sidebarAction:acceptItemsFromDropSession: is called. By default, the drop will be treated as a cancel operation if this is not implemented.

fn tabBarController_sidebar_sidebarAction_group_acceptItemsFromDropSession( &self, tab_bar_controller: &UITabBarController, sidebar: &UITabBarControllerSidebar, sidebar_action: &UIAction, group: &UITabGroup, session: &ProtocolObject<dyn UIDropSession>, )
where Self: Sized + Message,

Receive the drop from into the sidebarAction using the specified session. This is only called if the drop operation returned from tabBarController:sidebar:sidebarAction:operationForAcceptingItemsFromDropSession is valid for a drop.