objc2_ui_kit

Trait UITextFieldDelegate

pub unsafe trait UITextFieldDelegate: NSObjectProtocol + MainThreadOnly {

    // Provided methods
    unsafe fn textFieldShouldBeginEditing(
        &self,
        text_field: &UITextField,
    ) -> bool
       where Self: Sized + Message { ... }
    unsafe fn textFieldDidBeginEditing(&self, text_field: &UITextField)
       where Self: Sized + Message { ... }
    unsafe fn textFieldShouldEndEditing(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    unsafe fn textFieldDidEndEditing(&self, text_field: &UITextField)
       where Self: Sized + Message { ... }
    unsafe fn textFieldDidEndEditing_reason(
        &self,
        text_field: &UITextField,
        reason: UITextFieldDidEndEditingReason,
    )
       where Self: Sized + Message { ... }
    unsafe fn textField_shouldChangeCharactersInRange_replacementString(
        &self,
        text_field: &UITextField,
        range: NSRange,
        string: &NSString,
    ) -> bool
       where Self: Sized + Message { ... }
    unsafe fn textFieldDidChangeSelection(&self, text_field: &UITextField)
       where Self: Sized + Message { ... }
    unsafe fn textFieldShouldClear(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    unsafe fn textFieldShouldReturn(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    unsafe fn textField_editMenuForCharactersInRange_suggestedActions(
        &self,
        text_field: &UITextField,
        range: NSRange,
        suggested_actions: &NSArray<UIMenuElement>,
    ) -> Option<Retained<UIMenu>>
       where Self: Sized + Message { ... }
    unsafe fn textField_willPresentEditMenuWithAnimator(
        &self,
        text_field: &UITextField,
        animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>,
    )
       where Self: Sized + Message { ... }
    unsafe fn textField_willDismissEditMenuWithAnimator(
        &self,
        text_field: &UITextField,
        animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>,
    )
       where Self: Sized + Message { ... }
    unsafe fn textField_insertInputSuggestion(
        &self,
        text_field: &UITextField,
        input_suggestion: &UIInputSuggestion,
    )
       where Self: Sized + Message { ... }

}

Apple’s documentation

Provided Methods

unsafe fn textFieldShouldBeginEditing(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

unsafe fn textFieldDidBeginEditing(&self, text_field: &UITextField)
where Self: Sized + Message,

unsafe fn textFieldShouldEndEditing(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

unsafe fn textFieldDidEndEditing(&self, text_field: &UITextField)
where Self: Sized + Message,

unsafe fn textFieldDidEndEditing_reason( &self, text_field: &UITextField, reason: UITextFieldDidEndEditingReason, )
where Self: Sized + Message,

unsafe fn textField_shouldChangeCharactersInRange_replacementString( &self, text_field: &UITextField, range: NSRange, string: &NSString, ) -> bool
where Self: Sized + Message,

unsafe fn textFieldDidChangeSelection(&self, text_field: &UITextField)
where Self: Sized + Message,

unsafe fn textFieldShouldClear(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

unsafe fn textFieldShouldReturn(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

unsafe fn textField_editMenuForCharactersInRange_suggestedActions( &self, text_field: &UITextField, range: NSRange, suggested_actions: &NSArray<UIMenuElement>, ) -> Option<Retained<UIMenu>>
where Self: Sized + Message,

Asks the delegate for the menu to be shown for the specified text range.

Parameter textField: The text field requesting the menu.

Parameter range: The characters range for which the menu is presented for.

Parameter suggestedActions: The actions and commands that the system suggests.

Returns: Return a UIMenu describing the desired menu hierarchy. Return nilto present the default system menu.

unsafe fn textField_willPresentEditMenuWithAnimator( &self, text_field: &UITextField, animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>, )
where Self: Sized + Message,

Called when the text field is about to present the edit menu.

Parameter textField: The text field displaying the menu.

Parameter animator: Appearance animator. Add animations to this object to run them alongside the appearance transition.

unsafe fn textField_willDismissEditMenuWithAnimator( &self, text_field: &UITextField, animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>, )
where Self: Sized + Message,

Called when the text field is about to dismiss the edit menu.

Parameter textField: The text field displaying the menu.

Parameter animator: Dismissal animator. Add animations to this object to run them alongside the dismissal transition.

unsafe fn textField_insertInputSuggestion( &self, text_field: &UITextField, input_suggestion: &UIInputSuggestion, )
where Self: Sized + Message,

Tells the delegate when the keyboard delivers an input suggestion.

Parameters:

textField: The text field that is currently the first responder.

inputSuggestion: The input suggestion that the user or system selected.