objc2_ui_kit

Trait UITraitEnvironment

pub unsafe trait UITraitEnvironment: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn traitCollection(&self) -> Retained<UITraitCollection>
       where Self: Sized + Message { ... }
    fn traitCollectionDidChange(
        &self,
        previous_trait_collection: Option<&UITraitCollection>,
    )
       where Self: Sized + Message { ... }
}

Trait environments expose a trait collection that describes their environment.

See also Appleโs documentation

Provided Methods

fn traitCollection(&self) -> Retained<UITraitCollection>
where Self: Sized + Message,

fn traitCollectionDidChange( &self, previous_trait_collection: Option<&UITraitCollection>, )
where Self: Sized + Message,

๐Deprecated: Use the trait change registration APIs declared in the UITraitChangeObservable protocol

To be overridden as needed to provide custom behavior when the environmentโs traits change.