objc2_ui_kit

Trait UIViewControllerAnimatedTransitioning

pub unsafe trait UIViewControllerAnimatedTransitioning: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn transitionDuration(
        &self,
        transition_context: Option<&ProtocolObject<dyn UIViewControllerContextTransitioning>>,
    ) -> NSTimeInterval
       where Self: Sized + Message { ... }
    fn animateTransition(
        &self,
        transition_context: &ProtocolObject<dyn UIViewControllerContextTransitioning>,
    )
       where Self: Sized + Message { ... }
    fn interruptibleAnimatorForTransition(
        &self,
        transition_context: &ProtocolObject<dyn UIViewControllerContextTransitioning>,
    ) -> Retained<ProtocolObject<dyn UIViewImplicitlyAnimating>>
       where Self: Sized + Message { ... }
    fn animationEnded(&self, transition_completed: bool)
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn transitionDuration( &self, transition_context: Option<&ProtocolObject<dyn UIViewControllerContextTransitioning>>, ) -> NSTimeInterval
where Self: Sized + Message,

fn animateTransition( &self, transition_context: &ProtocolObject<dyn UIViewControllerContextTransitioning>, )
where Self: Sized + Message,

fn interruptibleAnimatorForTransition( &self, transition_context: &ProtocolObject<dyn UIViewControllerContextTransitioning>, ) -> Retained<ProtocolObject<dyn UIViewImplicitlyAnimating>>
where Self: Sized + Message,

A conforming object implements this method if the transition it creates can be interrupted. For example, it could return an instance of a UIViewPropertyAnimator. It is expected that this method will return the same instance for the life of a transition.

fn animationEnded(&self, transition_completed: bool)
where Self: Sized + Message,