objc2_ui_kit

Trait UICollectionViewDataSource

pub unsafe trait UICollectionViewDataSource: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn collectionView_numberOfItemsInSection(
        &self,
        collection_view: &UICollectionView,
        section: NSInteger,
    ) -> NSInteger
       where Self: Sized + Message { ... }
    fn collectionView_cellForItemAtIndexPath(
        &self,
        collection_view: &UICollectionView,
        index_path: &NSIndexPath,
    ) -> Retained<UICollectionViewCell>
       where Self: Sized + Message { ... }
    fn numberOfSectionsInCollectionView(
        &self,
        collection_view: &UICollectionView,
    ) -> NSInteger
       where Self: Sized + Message { ... }
    fn collectionView_viewForSupplementaryElementOfKind_atIndexPath(
        &self,
        collection_view: &UICollectionView,
        kind: &NSString,
        index_path: &NSIndexPath,
    ) -> Retained<UICollectionReusableView>
       where Self: Sized + Message { ... }
    fn collectionView_canMoveItemAtIndexPath(
        &self,
        collection_view: &UICollectionView,
        index_path: &NSIndexPath,
    ) -> bool
       where Self: Sized + Message { ... }
    fn collectionView_moveItemAtIndexPath_toIndexPath(
        &self,
        collection_view: &UICollectionView,
        source_index_path: &NSIndexPath,
        destination_index_path: &NSIndexPath,
    )
       where Self: Sized + Message { ... }
    fn indexTitlesForCollectionView(
        &self,
        collection_view: &UICollectionView,
    ) -> Option<Retained<NSArray<NSString>>>
       where Self: Sized + Message { ... }
    fn collectionView_indexPathForIndexTitle_atIndex(
        &self,
        collection_view: &UICollectionView,
        title: &NSString,
        index: NSInteger,
    ) -> Retained<NSIndexPath>
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn collectionView_numberOfItemsInSection( &self, collection_view: &UICollectionView, section: NSInteger, ) -> NSInteger
where Self: Sized + Message,

fn collectionView_cellForItemAtIndexPath( &self, collection_view: &UICollectionView, index_path: &NSIndexPath, ) -> Retained<UICollectionViewCell>
where Self: Sized + Message,

fn numberOfSectionsInCollectionView( &self, collection_view: &UICollectionView, ) -> NSInteger
where Self: Sized + Message,

fn collectionView_viewForSupplementaryElementOfKind_atIndexPath( &self, collection_view: &UICollectionView, kind: &NSString, index_path: &NSIndexPath, ) -> Retained<UICollectionReusableView>
where Self: Sized + Message,

fn collectionView_canMoveItemAtIndexPath( &self, collection_view: &UICollectionView, index_path: &NSIndexPath, ) -> bool
where Self: Sized + Message,

fn collectionView_moveItemAtIndexPath_toIndexPath( &self, collection_view: &UICollectionView, source_index_path: &NSIndexPath, destination_index_path: &NSIndexPath, )
where Self: Sized + Message,

fn indexTitlesForCollectionView( &self, collection_view: &UICollectionView, ) -> Option<Retained<NSArray<NSString>>>
where Self: Sized + Message,

Returns a list of index titles to display in the index view (e.g. [“A”, “B”, “C” … “Z”, “#”])

fn collectionView_indexPathForIndexTitle_atIndex( &self, collection_view: &UICollectionView, title: &NSString, index: NSInteger, ) -> Retained<NSIndexPath>
where Self: Sized + Message,

Returns the index path that corresponds to the given title / index. (e.g. “B”,1) Return an index path with a single index to indicate an entire section, instead of a specific item.