objc2_ui_kit

Trait UIMenuBuilder

pub unsafe trait UIMenuBuilder: MainThreadOnly {

    // Provided methods
    fn system(&self) -> Retained<UIMenuSystem>
       where Self: Sized + Message { ... }
    fn menuForIdentifier(
        &self,
        identifier: &UIMenuIdentifier,
    ) -> Option<Retained<UIMenu>>
       where Self: Sized + Message { ... }
    fn actionForIdentifier(
        &self,
        identifier: &UIActionIdentifier,
    ) -> Option<Retained<UIAction>>
       where Self: Sized + Message { ... }
    unsafe fn commandForAction_propertyList(
        &self,
        action: Sel,
        property_list: Option<&AnyObject>,
    ) -> Option<Retained<UICommand>>
       where Self: Sized + Message { ... }
    fn replaceMenuForIdentifier_withMenu(
        &self,
        replaced_identifier: &UIMenuIdentifier,
        replacement_menu: &UIMenu,
    )
       where Self: Sized + Message { ... }
    unsafe fn replaceChildrenOfMenuForIdentifier_fromChildrenBlock(
        &self,
        parent_identifier: &UIMenuIdentifier,
        children_block: &DynBlock<dyn Fn(NonNull<NSArray<UIMenuElement>>) -> NonNull<NSArray<UIMenuElement>> + '_>,
    )
       where Self: Sized + Message { ... }
    fn replaceMenuForIdentifier_withElements(
        &self,
        replaced_identifier: &UIMenuIdentifier,
        replacement_elements: &NSArray<UIMenuElement>,
    )
       where Self: Sized + Message { ... }
    fn replaceActionForIdentifier_withElements(
        &self,
        replaced_identifier: &UIActionIdentifier,
        replacement_elements: &NSArray<UIMenuElement>,
    )
       where Self: Sized + Message { ... }
    unsafe fn replaceCommandForAction_propertyList_withElements(
        &self,
        replaced_action: Sel,
        replaced_property_list: Option<&AnyObject>,
        replacement_elements: &NSArray<UIMenuElement>,
    )
       where Self: Sized + Message { ... }
    fn insertSiblingMenu_beforeMenuForIdentifier(
        &self,
        sibling_menu: &UIMenu,
        sibling_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertElements_beforeMenuForIdentifier(
        &self,
        inserted_elements: &NSArray<UIMenuElement>,
        sibling_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertElements_afterMenuForIdentifier(
        &self,
        inserted_elements: &NSArray<UIMenuElement>,
        sibling_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertSiblingMenu_afterMenuForIdentifier(
        &self,
        sibling_menu: &UIMenu,
        sibling_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertChildMenu_atStartOfMenuForIdentifier(
        &self,
        child_menu: &UIMenu,
        parent_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertElements_beforeActionForIdentifier(
        &self,
        inserted_elements: &NSArray<UIMenuElement>,
        sibling_identifier: &UIActionIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertElements_afterActionForIdentifier(
        &self,
        inserted_elements: &NSArray<UIMenuElement>,
        sibling_identifier: &UIActionIdentifier,
    )
       where Self: Sized + Message { ... }
    unsafe fn insertElements_beforeCommandForAction_propertyList(
        &self,
        inserted_elements: &NSArray<UIMenuElement>,
        sibling_action: Sel,
        sibling_property_list: Option<&AnyObject>,
    )
       where Self: Sized + Message { ... }
    unsafe fn insertElements_afterCommandForAction_propertyList(
        &self,
        inserted_elements: &NSArray<UIMenuElement>,
        sibling_action: Sel,
        sibling_property_list: Option<&AnyObject>,
    )
       where Self: Sized + Message { ... }
    fn insertElements_atStartOfMenuForIdentifier(
        &self,
        child_elements: &NSArray<UIMenuElement>,
        parent_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertChildMenu_atEndOfMenuForIdentifier(
        &self,
        child_menu: &UIMenu,
        parent_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn insertElements_atEndOfMenuForIdentifier(
        &self,
        child_elements: &NSArray<UIMenuElement>,
        parent_identifier: &UIMenuIdentifier,
    )
       where Self: Sized + Message { ... }
    fn removeMenuForIdentifier(&self, removed_identifier: &UIMenuIdentifier)
       where Self: Sized + Message { ... }
    fn removeActionForIdentifier(&self, removed_identifier: &UIActionIdentifier)
       where Self: Sized + Message { ... }
    unsafe fn removeCommandForAction_propertyList(
        &self,
        removed_action: Sel,
        removed_property_list: Option<&AnyObject>,
    )
       where Self: Sized + Message { ... }

}

Encapsulates access and mutation for a menu hierarchy.

See also Apple’s documentation

Provided Methods

fn system(&self) -> Retained<UIMenuSystem>
where Self: Sized + Message,

Which system we are building for.

fn menuForIdentifier( &self, identifier: &UIMenuIdentifier, ) -> Option<Retained<UIMenu>>
where Self: Sized + Message,

Fetch the identified menu.

Parameter identifier: The identifier of the menu to fetch.

Returns: The menu with the given identifier, or nil if no such menu.

fn actionForIdentifier( &self, identifier: &UIActionIdentifier, ) -> Option<Retained<UIAction>>
where Self: Sized + Message,

Fetch the identified action.

Parameter identifier: The identifier of the action to fetch.

Returns: The action with the given identifier, or nil if no such action.

unsafe fn commandForAction_propertyList( &self, action: Sel, property_list: Option<&AnyObject>, ) -> Option<Retained<UICommand>>
where Self: Sized + Message,

Fetch the identified command.

Parameter action: The action of the command to fetch.

Parameter propertyList: Property list object to distinguish commands, if needed.

Returns: The command with the given action and property list, or nil if no such command.

Safety

action must be a valid selector.

property_list should be of the correct type.

fn replaceMenuForIdentifier_withMenu( &self, replaced_identifier: &UIMenuIdentifier, replacement_menu: &UIMenu, )
where Self: Sized + Message,

Replace an identified menu with a menu.

Parameter replacedIdentifier: The identifier of the menu to be replaced.

Parameter replacementGroup: The replacement menu.

unsafe fn replaceChildrenOfMenuForIdentifier_fromChildrenBlock( &self, parent_identifier: &UIMenuIdentifier, children_block: &DynBlock<dyn Fn(NonNull<NSArray<UIMenuElement>>) -> NonNull<NSArray<UIMenuElement>> + '_>, )
where Self: Sized + Message,

Replace the children of an identified parent menu.

Parameter parentIdentifier: The identifier of the parent menu.

Parameter childrenBlock: A block that returns the new children, given the old children.

Safety

children_block block’s return must be a valid pointer.

fn replaceMenuForIdentifier_withElements( &self, replaced_identifier: &UIMenuIdentifier, replacement_elements: &NSArray<UIMenuElement>, )
where Self: Sized + Message,

Replace an identified menu with menu elements.

Parameter replacedIdentifier: The identifier of the menu to be replaced.

Parameter replacementElements: The replacement elements.

fn replaceActionForIdentifier_withElements( &self, replaced_identifier: &UIActionIdentifier, replacement_elements: &NSArray<UIMenuElement>, )
where Self: Sized + Message,

Replace an identified action with menu elements.

Parameter replacedIdentifier: The identifier of the action to be replaced.

Parameter replacementElements: The replacement elements.

unsafe fn replaceCommandForAction_propertyList_withElements( &self, replaced_action: Sel, replaced_property_list: Option<&AnyObject>, replacement_elements: &NSArray<UIMenuElement>, )
where Self: Sized + Message,

Replace an identified command with menu elements.

Parameter replacedAction: The action of the command to be replaced.

Parameter replacedPropertyList: Property list object to distinguish commands, if needed.

Parameter replacementElements: The replacement elements.

Safety

replaced_action must be a valid selector.

replaced_property_list should be of the correct type.

fn insertSiblingMenu_beforeMenuForIdentifier( &self, sibling_menu: &UIMenu, sibling_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert a sibling menu before an identified sibling menu.

Parameter siblingGroup: The sibling menu to insert.

Parameter siblingIdentifier: The identifier of the sibling menu to insert before.

fn insertElements_beforeMenuForIdentifier( &self, inserted_elements: &NSArray<UIMenuElement>, sibling_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert elements before an identified menu.

Parameter insertedElements: The elements to insert.

Parameter siblingIdentifier: The identifier of the menu to insert elements before.

fn insertElements_afterMenuForIdentifier( &self, inserted_elements: &NSArray<UIMenuElement>, sibling_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert elements after an identified menu.

Parameter insertedElements: The elements to insert.

Parameter siblingIdentifier: The identifier of the menu to insert elements after.

fn insertSiblingMenu_afterMenuForIdentifier( &self, sibling_menu: &UIMenu, sibling_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert a sibling menu after an identified sibling menu.

Parameter siblingGroup: The sibling menu to insert.

Parameter siblingIdentifier: The identifier of the sibling menu to insert after.

fn insertChildMenu_atStartOfMenuForIdentifier( &self, child_menu: &UIMenu, parent_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert a child menu at the start of an identified parent menu.

Parameter childGroup: The child menu to insert.

Parameter parentIdentifier: The identifier of the parent menu to insert at the start of.

fn insertElements_beforeActionForIdentifier( &self, inserted_elements: &NSArray<UIMenuElement>, sibling_identifier: &UIActionIdentifier, )
where Self: Sized + Message,

Insert elements before an identified action.

Parameter insertedElements: The elements to insert.

Parameter siblingIdentifier: The identifier of the action to insert elements before.

fn insertElements_afterActionForIdentifier( &self, inserted_elements: &NSArray<UIMenuElement>, sibling_identifier: &UIActionIdentifier, )
where Self: Sized + Message,

Insert elements after an identified action.

Parameter insertedElements: The elements to insert.

Parameter siblingIdentifier: The identifier of the action to insert elements after.

unsafe fn insertElements_beforeCommandForAction_propertyList( &self, inserted_elements: &NSArray<UIMenuElement>, sibling_action: Sel, sibling_property_list: Option<&AnyObject>, )
where Self: Sized + Message,

Insert elements before an identified command.

Parameter insertedElements: The elements to insert.

Parameter siblingAction: The action of the command to insert elements before.

Parameter siblingPropertyList: Property list object to distinguish commands, if needed.

Safety

sibling_action must be a valid selector.

sibling_property_list should be of the correct type.

unsafe fn insertElements_afterCommandForAction_propertyList( &self, inserted_elements: &NSArray<UIMenuElement>, sibling_action: Sel, sibling_property_list: Option<&AnyObject>, )
where Self: Sized + Message,

Insert elements after an identified command.

Parameter insertedElements: The elements to insert.

Parameter siblingAction: The action of the command to insert elements after.

Parameter siblingPropertyList: Property list object to distinguish commands, if needed.

Safety

sibling_action must be a valid selector.

sibling_property_list should be of the correct type.

fn insertElements_atStartOfMenuForIdentifier( &self, child_elements: &NSArray<UIMenuElement>, parent_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert elements at the start of an identified parent menu.

Parameter childElements: The child elements to insert.

Parameter parentIdentifier: The identifier of the parent menu to insert elements at the start of.

fn insertChildMenu_atEndOfMenuForIdentifier( &self, child_menu: &UIMenu, parent_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert a child menu at the end of an identified parent menu.

Parameter childGroup: The child menu to insert.

Parameter parentIdentifier: The identifier of the parent menu to insert at the end of.

fn insertElements_atEndOfMenuForIdentifier( &self, child_elements: &NSArray<UIMenuElement>, parent_identifier: &UIMenuIdentifier, )
where Self: Sized + Message,

Insert elements at the end of an identified parent menu.

Parameter childElements: The child elements to insert.

Parameter parentIdentifier: The identifier of the parent menu to insert elements at the end of.

fn removeMenuForIdentifier(&self, removed_identifier: &UIMenuIdentifier)
where Self: Sized + Message,

Remove an identified menu.

Parameter removedIdentifier: The menu to remove.

fn removeActionForIdentifier(&self, removed_identifier: &UIActionIdentifier)
where Self: Sized + Message,

Remove an identified action.

Parameter removedIdentifier: The identifier of the action to remove.

unsafe fn removeCommandForAction_propertyList( &self, removed_action: Sel, removed_property_list: Option<&AnyObject>, )
where Self: Sized + Message,

Remove an identified command.

Parameter removedAction: The action of the command to remove.

Parameter removedPropertyList: Property list object to distinguish commands, if needed.

Safety

removed_action must be a valid selector.

removed_property_list should be of the correct type.