objc2_ui_kit

Trait UIPencilInteractionDelegate

pub unsafe trait UIPencilInteractionDelegate: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn pencilInteractionDidTap(&self, interaction: &UIPencilInteraction)
       where Self: Sized + Message { ... }
    fn pencilInteraction_didReceiveTap(
        &self,
        interaction: &UIPencilInteraction,
        tap: &UIPencilInteractionTap,
    )
       where Self: Sized + Message { ... }
    fn pencilInteraction_didReceiveSqueeze(
        &self,
        interaction: &UIPencilInteraction,
        squeeze: &UIPencilInteractionSqueeze,
    )
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn pencilInteractionDidTap(&self, interaction: &UIPencilInteraction)
where Self: Sized + Message,

👎Deprecated: Use pencilInteraction(_:didReceiveTap:) instead

Called when the user taps on the side of the pencil if the interaction’s view is in a visible view hierarchy.

The delegate may do anything in this method. To perform the user’s preferred tap action (as set in Settings), consult the preferredTapAction class property of UIPencilInteraction.

Note: If pencilInteraction(_:didReceiveTap:) is implemented, this method will not be called.

fn pencilInteraction_didReceiveTap( &self, interaction: &UIPencilInteraction, tap: &UIPencilInteractionTap, )
where Self: Sized + Message,

Called when the user taps on the side of the pencil if the interaction’s view is in a visible view hierarchy.

The delegate may do anything in this method. To perform the user’s preferred tap action (as set in Settings), consult the preferredTapAction class property of UIPencilInteraction.

Note: If both this method and the deprecated pencilInteractionDidTap(_:) are implemented, then only this method will be called.

fn pencilInteraction_didReceiveSqueeze( &self, interaction: &UIPencilInteraction, squeeze: &UIPencilInteractionSqueeze, )
where Self: Sized + Message,

Called when the user squeezes side of the pencil if the interaction’s view is in a visible view hierarchy.

The delegate may do anything in this method. To perform the user’s preferred squeeze action (as set in Settings), consult the preferredSqueezeAction class property of UIPencilInteraction.