objc2_ui_kit

Trait UISearchSuggestion

pub unsafe trait UISearchSuggestion: NSObjectProtocol + MainThreadOnly {
    // Provided methods
    fn localizedSuggestion(&self) -> Option<Retained<NSString>>
       where Self: Sized + Message { ... }
    fn localizedDescription(&self) -> Option<Retained<NSString>>
       where Self: Sized + Message { ... }
    fn iconImage(&self) -> Option<Retained<UIImage>>
       where Self: Sized + Message { ... }
    fn localizedAttributedSuggestion(
        &self,
    ) -> Option<Retained<NSAttributedString>>
       where Self: Sized + Message { ... }
    fn representedObject(&self) -> Option<Retained<AnyObject>>
       where Self: Sized + Message { ... }
    unsafe fn setRepresentedObject(
        &self,
        represented_object: Option<&AnyObject>,
    )
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn localizedSuggestion(&self) -> Option<Retained<NSString>>
where Self: Sized + Message,

The localized suggestion that will be displayed as the search string

fn localizedDescription(&self) -> Option<Retained<NSString>>
where Self: Sized + Message,

The localized description used as accessibility string

fn iconImage(&self) -> Option<Retained<UIImage>>
where Self: Sized + Message,

Icon that represents this search suggestion item

fn localizedAttributedSuggestion(&self) -> Option<Retained<NSAttributedString>>
where Self: Sized + Message,

If non-null, will be displayed as the search string instead of localizedSuggestion

fn representedObject(&self) -> Option<Retained<AnyObject>>
where Self: Sized + Message,

User information

unsafe fn setRepresentedObject(&self, represented_object: Option<&AnyObject>)
where Self: Sized + Message,

Setter for representedObject.

Safety

represented_object should be of the correct type.