objc2_ui_kit

Trait UITextFieldDelegate

pub unsafe trait UITextFieldDelegate: NSObjectProtocol + MainThreadOnly {

    // Provided methods
    fn textFieldShouldBeginEditing(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    fn textFieldDidBeginEditing(&self, text_field: &UITextField)
       where Self: Sized + Message { ... }
    fn textFieldShouldEndEditing(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    fn textFieldDidEndEditing(&self, text_field: &UITextField)
       where Self: Sized + Message { ... }
    fn textFieldDidEndEditing_reason(
        &self,
        text_field: &UITextField,
        reason: UITextFieldDidEndEditingReason,
    )
       where Self: Sized + Message { ... }
    fn textField_shouldChangeCharactersInRange_replacementString(
        &self,
        text_field: &UITextField,
        range: NSRange,
        string: &NSString,
    ) -> bool
       where Self: Sized + Message { ... }
    fn textField_shouldChangeCharactersInRanges_replacementString(
        &self,
        text_field: &UITextField,
        ranges: &NSArray<NSValue>,
        string: &NSString,
    ) -> bool
       where Self: Sized + Message { ... }
    fn textFieldDidChangeSelection(&self, text_field: &UITextField)
       where Self: Sized + Message { ... }
    fn textFieldShouldClear(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    fn textFieldShouldReturn(&self, text_field: &UITextField) -> bool
       where Self: Sized + Message { ... }
    fn textField_editMenuForCharactersInRange_suggestedActions(
        &self,
        text_field: &UITextField,
        range: NSRange,
        suggested_actions: &NSArray<UIMenuElement>,
    ) -> Option<Retained<UIMenu>>
       where Self: Sized + Message { ... }
    fn textField_editMenuForCharactersInRanges_suggestedActions(
        &self,
        text_field: &UITextField,
        ranges: &NSArray<NSValue>,
        suggested_actions: &NSArray<UIMenuElement>,
    ) -> Option<Retained<UIMenu>>
       where Self: Sized + Message { ... }
    fn textField_willPresentEditMenuWithAnimator(
        &self,
        text_field: &UITextField,
        animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>,
    )
       where Self: Sized + Message { ... }
    fn textField_willDismissEditMenuWithAnimator(
        &self,
        text_field: &UITextField,
        animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>,
    )
       where Self: Sized + Message { ... }
    fn textField_insertInputSuggestion(
        &self,
        text_field: &UITextField,
        input_suggestion: &UIInputSuggestion,
    )
       where Self: Sized + Message { ... }

}

Apple’s documentation

Provided Methods

fn textFieldShouldBeginEditing(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

fn textFieldDidBeginEditing(&self, text_field: &UITextField)
where Self: Sized + Message,

fn textFieldShouldEndEditing(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

fn textFieldDidEndEditing(&self, text_field: &UITextField)
where Self: Sized + Message,

fn textFieldDidEndEditing_reason( &self, text_field: &UITextField, reason: UITextFieldDidEndEditingReason, )
where Self: Sized + Message,

fn textField_shouldChangeCharactersInRange_replacementString( &self, text_field: &UITextField, range: NSRange, string: &NSString, ) -> bool
where Self: Sized + Message,

👎Deprecated

fn textField_shouldChangeCharactersInRanges_replacementString( &self, text_field: &UITextField, ranges: &NSArray<NSValue>, string: &NSString, ) -> bool
where Self: Sized + Message,

Asks the delegate if the text at the specified ranges should be replaced with string.

If this method returns YES then the text field will, at its own discretion, choose any one of the specified ranges of text and replace it with the specified replacementString before deleting the text at the other ranges.

Parameter textField: The text field asking the delegate

Parameter ranges: The ranges of the text that should be deleted before replacing

Parameter replacementString: The replacement string

Returns: Returns YES if the text at the ranges should be replaced.

fn textFieldDidChangeSelection(&self, text_field: &UITextField)
where Self: Sized + Message,

fn textFieldShouldClear(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

fn textFieldShouldReturn(&self, text_field: &UITextField) -> bool
where Self: Sized + Message,

fn textField_editMenuForCharactersInRange_suggestedActions( &self, text_field: &UITextField, range: NSRange, suggested_actions: &NSArray<UIMenuElement>, ) -> Option<Retained<UIMenu>>
where Self: Sized + Message,

👎Deprecated

Asks the delegate for the menu to be shown for the specified text range.

Parameter textField: The text field requesting the menu.

Parameter range: The characters range for which the menu is presented for.

Parameter suggestedActions: The actions and commands that the system suggests.

Returns: Return a UIMenu describing the desired menu hierarchy. Return nilto present the default system menu.

fn textField_editMenuForCharactersInRanges_suggestedActions( &self, text_field: &UITextField, ranges: &NSArray<NSValue>, suggested_actions: &NSArray<UIMenuElement>, ) -> Option<Retained<UIMenu>>
where Self: Sized + Message,

Asks the delegate for the menu to be shown for the specified ranges.

Parameter textField: The text field requesting the menu.

Parameter ranges: The text ranges for which the menu is presented for.

Parameter suggestedActions: The actions and commands that the system suggests.

Returns: Return a UIMenu describing the desired menu hierarchy. Return nilto present the default system menu.

fn textField_willPresentEditMenuWithAnimator( &self, text_field: &UITextField, animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>, )
where Self: Sized + Message,

Called when the text field is about to present the edit menu.

Parameter textField: The text field displaying the menu.

Parameter animator: Appearance animator. Add animations to this object to run them alongside the appearance transition.

fn textField_willDismissEditMenuWithAnimator( &self, text_field: &UITextField, animator: &ProtocolObject<dyn UIEditMenuInteractionAnimating>, )
where Self: Sized + Message,

Called when the text field is about to dismiss the edit menu.

Parameter textField: The text field displaying the menu.

Parameter animator: Dismissal animator. Add animations to this object to run them alongside the dismissal transition.

fn textField_insertInputSuggestion( &self, text_field: &UITextField, input_suggestion: &UIInputSuggestion, )
where Self: Sized + Message,

Tells the delegate when the keyboard delivers an input suggestion.

Parameters:

textField: The text field that is currently the first responder.

inputSuggestion: The input suggestion that the user or system selected.