objc2_ui_kit

Trait UITextInputTraits

pub unsafe trait UITextInputTraits: NSObjectProtocol + MainThreadOnly {

    // Provided methods
    fn autocapitalizationType(&self) -> UITextAutocapitalizationType
       where Self: Sized + Message { ... }
    fn setAutocapitalizationType(
        &self,
        autocapitalization_type: UITextAutocapitalizationType,
    )
       where Self: Sized + Message { ... }
    fn autocorrectionType(&self) -> UITextAutocorrectionType
       where Self: Sized + Message { ... }
    fn setAutocorrectionType(
        &self,
        autocorrection_type: UITextAutocorrectionType,
    )
       where Self: Sized + Message { ... }
    fn spellCheckingType(&self) -> UITextSpellCheckingType
       where Self: Sized + Message { ... }
    fn setSpellCheckingType(&self, spell_checking_type: UITextSpellCheckingType)
       where Self: Sized + Message { ... }
    fn smartQuotesType(&self) -> UITextSmartQuotesType
       where Self: Sized + Message { ... }
    fn setSmartQuotesType(&self, smart_quotes_type: UITextSmartQuotesType)
       where Self: Sized + Message { ... }
    fn smartDashesType(&self) -> UITextSmartDashesType
       where Self: Sized + Message { ... }
    fn setSmartDashesType(&self, smart_dashes_type: UITextSmartDashesType)
       where Self: Sized + Message { ... }
    fn smartInsertDeleteType(&self) -> UITextSmartInsertDeleteType
       where Self: Sized + Message { ... }
    fn setSmartInsertDeleteType(
        &self,
        smart_insert_delete_type: UITextSmartInsertDeleteType,
    )
       where Self: Sized + Message { ... }
    fn inlinePredictionType(&self) -> UITextInlinePredictionType
       where Self: Sized + Message { ... }
    fn setInlinePredictionType(
        &self,
        inline_prediction_type: UITextInlinePredictionType,
    )
       where Self: Sized + Message { ... }
    fn mathExpressionCompletionType(&self) -> UITextMathExpressionCompletionType
       where Self: Sized + Message { ... }
    fn setMathExpressionCompletionType(
        &self,
        math_expression_completion_type: UITextMathExpressionCompletionType,
    )
       where Self: Sized + Message { ... }
    fn keyboardType(&self) -> UIKeyboardType
       where Self: Sized + Message { ... }
    fn setKeyboardType(&self, keyboard_type: UIKeyboardType)
       where Self: Sized + Message { ... }
    fn keyboardAppearance(&self) -> UIKeyboardAppearance
       where Self: Sized + Message { ... }
    fn setKeyboardAppearance(&self, keyboard_appearance: UIKeyboardAppearance)
       where Self: Sized + Message { ... }
    fn returnKeyType(&self) -> UIReturnKeyType
       where Self: Sized + Message { ... }
    fn setReturnKeyType(&self, return_key_type: UIReturnKeyType)
       where Self: Sized + Message { ... }
    fn enablesReturnKeyAutomatically(&self) -> bool
       where Self: Sized + Message { ... }
    fn setEnablesReturnKeyAutomatically(
        &self,
        enables_return_key_automatically: bool,
    )
       where Self: Sized + Message { ... }
    fn isSecureTextEntry(&self) -> bool
       where Self: Sized + Message { ... }
    fn setSecureTextEntry(&self, secure_text_entry: bool)
       where Self: Sized + Message { ... }
    fn textContentType(&self) -> Retained<UITextContentType>
       where Self: Sized + Message { ... }
    fn setTextContentType(&self, text_content_type: Option<&UITextContentType>)
       where Self: Sized + Message { ... }
    fn passwordRules(&self) -> Option<Retained<UITextInputPasswordRules>>
       where Self: Sized + Message { ... }
    fn setPasswordRules(
        &self,
        password_rules: Option<&UITextInputPasswordRules>,
    )
       where Self: Sized + Message { ... }
    fn writingToolsBehavior(&self) -> UIWritingToolsBehavior
       where Self: Sized + Message { ... }
    fn setWritingToolsBehavior(
        &self,
        writing_tools_behavior: UIWritingToolsBehavior,
    )
       where Self: Sized + Message { ... }
    fn allowedWritingToolsResultOptions(&self) -> UIWritingToolsResultOptions
       where Self: Sized + Message { ... }
    fn setAllowedWritingToolsResultOptions(
        &self,
        allowed_writing_tools_result_options: UIWritingToolsResultOptions,
    )
       where Self: Sized + Message { ... }
    fn conversationContext(&self) -> Option<Retained<UIConversationContext>>
       where Self: Sized + Message { ... }
    fn setConversationContext(
        &self,
        conversation_context: Option<&UIConversationContext>,
    )
       where Self: Sized + Message { ... }
    fn allowsNumberPadPopover(&self) -> bool
       where Self: Sized + Message { ... }
    fn setAllowsNumberPadPopover(&self, allows_number_pad_popover: bool)
       where Self: Sized + Message { ... }

}

Apple’s documentation

Provided Methods

fn autocapitalizationType(&self) -> UITextAutocapitalizationType
where Self: Sized + Message,

fn setAutocapitalizationType( &self, autocapitalization_type: UITextAutocapitalizationType, )
where Self: Sized + Message,

Setter for autocapitalizationType.

fn autocorrectionType(&self) -> UITextAutocorrectionType
where Self: Sized + Message,

fn setAutocorrectionType(&self, autocorrection_type: UITextAutocorrectionType)
where Self: Sized + Message,

Setter for autocorrectionType.

fn spellCheckingType(&self) -> UITextSpellCheckingType
where Self: Sized + Message,

fn setSpellCheckingType(&self, spell_checking_type: UITextSpellCheckingType)
where Self: Sized + Message,

Setter for spellCheckingType.

fn smartQuotesType(&self) -> UITextSmartQuotesType
where Self: Sized + Message,

fn setSmartQuotesType(&self, smart_quotes_type: UITextSmartQuotesType)
where Self: Sized + Message,

Setter for smartQuotesType.

fn smartDashesType(&self) -> UITextSmartDashesType
where Self: Sized + Message,

fn setSmartDashesType(&self, smart_dashes_type: UITextSmartDashesType)
where Self: Sized + Message,

Setter for smartDashesType.

fn smartInsertDeleteType(&self) -> UITextSmartInsertDeleteType
where Self: Sized + Message,

fn setSmartInsertDeleteType( &self, smart_insert_delete_type: UITextSmartInsertDeleteType, )
where Self: Sized + Message,

Setter for smartInsertDeleteType.

fn inlinePredictionType(&self) -> UITextInlinePredictionType
where Self: Sized + Message,

fn setInlinePredictionType( &self, inline_prediction_type: UITextInlinePredictionType, )
where Self: Sized + Message,

Setter for inlinePredictionType.

fn mathExpressionCompletionType(&self) -> UITextMathExpressionCompletionType
where Self: Sized + Message,

fn setMathExpressionCompletionType( &self, math_expression_completion_type: UITextMathExpressionCompletionType, )
where Self: Sized + Message,

Setter for mathExpressionCompletionType.

fn keyboardType(&self) -> UIKeyboardType
where Self: Sized + Message,

fn setKeyboardType(&self, keyboard_type: UIKeyboardType)
where Self: Sized + Message,

Setter for keyboardType.

fn keyboardAppearance(&self) -> UIKeyboardAppearance
where Self: Sized + Message,

fn setKeyboardAppearance(&self, keyboard_appearance: UIKeyboardAppearance)
where Self: Sized + Message,

Setter for keyboardAppearance.

fn returnKeyType(&self) -> UIReturnKeyType
where Self: Sized + Message,

fn setReturnKeyType(&self, return_key_type: UIReturnKeyType)
where Self: Sized + Message,

Setter for returnKeyType.

fn enablesReturnKeyAutomatically(&self) -> bool
where Self: Sized + Message,

fn setEnablesReturnKeyAutomatically( &self, enables_return_key_automatically: bool, )
where Self: Sized + Message,

Setter for enablesReturnKeyAutomatically.

fn isSecureTextEntry(&self) -> bool
where Self: Sized + Message,

fn setSecureTextEntry(&self, secure_text_entry: bool)
where Self: Sized + Message,

Setter for isSecureTextEntry.

fn textContentType(&self) -> Retained<UITextContentType>
where Self: Sized + Message,

fn setTextContentType(&self, text_content_type: Option<&UITextContentType>)
where Self: Sized + Message,

Setter for textContentType.

This is copied when set.

fn passwordRules(&self) -> Option<Retained<UITextInputPasswordRules>>
where Self: Sized + Message,

fn setPasswordRules(&self, password_rules: Option<&UITextInputPasswordRules>)
where Self: Sized + Message,

Setter for passwordRules.

This is copied when set.

fn writingToolsBehavior(&self) -> UIWritingToolsBehavior
where Self: Sized + Message,

fn setWritingToolsBehavior( &self, writing_tools_behavior: UIWritingToolsBehavior, )
where Self: Sized + Message,

Setter for writingToolsBehavior.

fn allowedWritingToolsResultOptions(&self) -> UIWritingToolsResultOptions
where Self: Sized + Message,

fn setAllowedWritingToolsResultOptions( &self, allowed_writing_tools_result_options: UIWritingToolsResultOptions, )
where Self: Sized + Message,

Setter for allowedWritingToolsResultOptions.

fn conversationContext(&self) -> Option<Retained<UIConversationContext>>
where Self: Sized + Message,

A reference to a conversation, such as a mail or messaging thread.

Set this conversation context before the keyboard appears; the keyboard uses this context to initialize its conversation context value. When updates occur in the conversation, call UITextInputDelegate/conversationContext(_:didChange:) on the inputDelegate property for UITextInput objects, such as UITextView/inputDelegateorUITextField/inputDelegate``.

fn setConversationContext( &self, conversation_context: Option<&UIConversationContext>, )
where Self: Sized + Message,

Setter for conversationContext.

fn allowsNumberPadPopover(&self) -> bool
where Self: Sized + Message,

fn setAllowsNumberPadPopover(&self, allows_number_pad_popover: bool)
where Self: Sized + Message,

Setter for allowsNumberPadPopover.