objc2_ui_kit

Trait UIViewControllerTransitionCoordinatorContext

pub unsafe trait UIViewControllerTransitionCoordinatorContext: NSObjectProtocol + MainThreadOnly {

    // Provided methods
    fn isAnimated(&self) -> bool
       where Self: Sized + Message { ... }
    fn presentationStyle(&self) -> UIModalPresentationStyle
       where Self: Sized + Message { ... }
    fn initiallyInteractive(&self) -> bool
       where Self: Sized + Message { ... }
    fn isInterruptible(&self) -> bool
       where Self: Sized + Message { ... }
    fn isInteractive(&self) -> bool
       where Self: Sized + Message { ... }
    fn isCancelled(&self) -> bool
       where Self: Sized + Message { ... }
    fn transitionDuration(&self) -> NSTimeInterval
       where Self: Sized + Message { ... }
    fn percentComplete(&self) -> CGFloat
       where Self: Sized + Message { ... }
    fn completionVelocity(&self) -> CGFloat
       where Self: Sized + Message { ... }
    fn completionCurve(&self) -> UIViewAnimationCurve
       where Self: Sized + Message { ... }
    fn viewControllerForKey(
        &self,
        key: &UITransitionContextViewControllerKey,
    ) -> Option<Retained<UIViewController>>
       where Self: Sized + Message { ... }
    fn viewForKey(
        &self,
        key: &UITransitionContextViewKey,
    ) -> Option<Retained<UIView>>
       where Self: Sized + Message { ... }
    fn containerView(&self) -> Retained<UIView>
       where Self: Sized + Message { ... }
    fn targetTransform(&self) -> CGAffineTransform
       where Self: Sized + Message { ... }

}

Apple’s documentation

Provided Methods

fn isAnimated(&self) -> bool
where Self: Sized + Message,

fn presentationStyle(&self) -> UIModalPresentationStyle
where Self: Sized + Message,

fn initiallyInteractive(&self) -> bool
where Self: Sized + Message,

initiallyInteractive indicates whether the transition was initiated as an interactive transition. It never changes during the course of a transition. It can only be YES if isAnimated is YES. If it is NO, then isInteractive can only be YES if isInterruptible is YES

fn isInterruptible(&self) -> bool
where Self: Sized + Message,

fn isInteractive(&self) -> bool
where Self: Sized + Message,

fn isCancelled(&self) -> bool
where Self: Sized + Message,

fn transitionDuration(&self) -> NSTimeInterval
where Self: Sized + Message,

fn percentComplete(&self) -> CGFloat
where Self: Sized + Message,

fn completionVelocity(&self) -> CGFloat
where Self: Sized + Message,

fn completionCurve(&self) -> UIViewAnimationCurve
where Self: Sized + Message,

fn viewControllerForKey( &self, key: &UITransitionContextViewControllerKey, ) -> Option<Retained<UIViewController>>
where Self: Sized + Message,

fn viewForKey( &self, key: &UITransitionContextViewKey, ) -> Option<Retained<UIView>>
where Self: Sized + Message,

fn containerView(&self) -> Retained<UIView>
where Self: Sized + Message,

fn targetTransform(&self) -> CGAffineTransform
where Self: Sized + Message,