objc2_ui_kit

Trait UIWindowSceneDelegate

pub unsafe trait UIWindowSceneDelegate: UISceneDelegate + MainThreadOnly {
    // Provided methods
    fn window(&self) -> Option<Retained<UIWindow>>
       where Self: Sized + Message { ... }
    fn setWindow(&self, window: Option<&UIWindow>)
       where Self: Sized + Message { ... }
    fn windowScene_didUpdateCoordinateSpace_interfaceOrientation_traitCollection(
        &self,
        window_scene: &UIWindowScene,
        previous_coordinate_space: &ProtocolObject<dyn UICoordinateSpace>,
        previous_interface_orientation: UIInterfaceOrientation,
        previous_trait_collection: &UITraitCollection,
    )
       where Self: Sized + Message { ... }
    fn windowScene_didUpdateEffectiveGeometry(
        &self,
        window_scene: &UIWindowScene,
        previous_effective_geometry: &UIWindowSceneGeometry,
    )
       where Self: Sized + Message { ... }
    fn windowScene_performActionForShortcutItem_completionHandler(
        &self,
        window_scene: &UIWindowScene,
        shortcut_item: &UIApplicationShortcutItem,
        completion_handler: &DynBlock<dyn Fn(Bool)>,
    )
       where Self: Sized + Message { ... }
    fn windowScene_userDidAcceptCloudKitShareWithMetadata(
        &self,
        window_scene: &UIWindowScene,
        cloud_kit_share_metadata: &CKShareMetadata,
    )
       where Self: Sized + Message { ... }
    fn preferredWindowingControlStyleForScene(
        &self,
        window_scene: &UIWindowScene,
    ) -> Retained<UISceneWindowingControlStyle>
       where Self: Sized + Message { ... }
}

Apple’s documentation

Provided Methods

fn window(&self) -> Option<Retained<UIWindow>>
where Self: Sized + Message,

fn setWindow(&self, window: Option<&UIWindow>)
where Self: Sized + Message,

Setter for window.

fn windowScene_didUpdateCoordinateSpace_interfaceOrientation_traitCollection( &self, window_scene: &UIWindowScene, previous_coordinate_space: &ProtocolObject<dyn UICoordinateSpace>, previous_interface_orientation: UIInterfaceOrientation, previous_trait_collection: &UITraitCollection, )
where Self: Sized + Message,

👎Deprecated: Use windowScene(_: didUpdateEffectiveGeometry:) to be notified of the scene’s geometry changes, or use traits whose values are inherited from the scene via the traitCollection of views and view controllers instead.

Called when the coordinate space, interface orientation, or trait collection of a UIWindowScene changes.

Always called when a UIWindowScene moves between screens.

fn windowScene_didUpdateEffectiveGeometry( &self, window_scene: &UIWindowScene, previous_effective_geometry: &UIWindowSceneGeometry, )
where Self: Sized + Message,

Called when the window scene’s effective geometry has changed.

Always called when a UIWindowScene moves between screens.

fn windowScene_performActionForShortcutItem_completionHandler( &self, window_scene: &UIWindowScene, shortcut_item: &UIApplicationShortcutItem, completion_handler: &DynBlock<dyn Fn(Bool)>, )
where Self: Sized + Message,

fn windowScene_userDidAcceptCloudKitShareWithMetadata( &self, window_scene: &UIWindowScene, cloud_kit_share_metadata: &CKShareMetadata, )
where Self: Sized + Message,

fn preferredWindowingControlStyleForScene( &self, window_scene: &UIWindowScene, ) -> Retained<UISceneWindowingControlStyle>
where Self: Sized + Message,

Called by the system to determine the windowing control style for the provided scene. automaticStyle will be used if this method is not implemented.